% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{addIOVToOmegaMatrix}
\alias{addIOVToOmegaMatrix}
\title{Add variance-covariance information regarding all IOV terms in the omega matrix.
Note that the column and row names of the returned matrix are strictly identical to the ones returned by getParameterNames().}
\usage{
addIOVToOmegaMatrix(tdmore, occasions)
}
\arguments{
\item{tdmore}{the tdmore object}

\item{occasions}{how many occasions}
}
\value{
the omega matrix with the duplicated IOV terms at the end
}
\description{
Add variance-covariance information regarding all IOV terms in the omega matrix.
Note that the column and row names of the returned matrix are strictly identical to the ones returned by getParameterNames().
}
