% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc.R
\name{addIterationColumn}
\alias{addIterationColumn}
\title{Add an iteration column 'ITER' to the observed dataframe.
The iteration column determines in which MPC iteration each observation
will be included.}
\usage{
addIterationColumn(regimen, observed)
}
\arguments{
\item{regimen}{regimen}

\item{observed}{observed}
}
\value{
modified observed dataframe, with an 'ITER' column
}
\description{
This is calculated by matching the corresponding OCC from the regimen (time-wise).
}
