% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{as.population}
\alias{as.population}
\title{Transform a tdmore fit object into a population fit
This is done by re-executing the 'estimate' function, but with no observed data}
\usage{
as.population(x)
}
\arguments{
\item{x}{a tdmorefit object}
}
\description{
Transform a tdmore fit object into a population fit
This is done by re-executing the 'estimate' function, but with no observed data
}
