% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{as.RxODE_regimen}
\alias{as.RxODE_regimen}
\title{Add the given regimen into the RxODE event table.}
\usage{
as.RxODE_regimen(regimen, nbSSDoses = NULL)
}
\arguments{
\item{regimen}{the specified regimen}

\item{nbSSDoses}{number of doses to prepend for simulation of steady-state}
}
\value{
a completed event table
}
\description{
Add the given regimen into the RxODE event table.
}
