% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{as.sample}
\alias{as.sample}
\title{Create a tiible of N random tdmorefit instances}
\usage{
as.sample(x, N = 100)
}
\arguments{
\item{x}{a tdmorefit or tdmore object}

\item{N}{the number of random subjects to create}
}
\description{
Create a tiible of N random tdmorefit instances
}
\details{
This method samples ETA parameters from the variance-covariance
matrix of a given tdmorefit object. If a tdmore object is provided,
the population is used.

The regimen needs to be provided to know the number of a
}
