% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.autolayer.R
\name{autolayer.recommendation}
\alias{autolayer.recommendation}
\title{Automatically create the required layers for a recommendation object}
\usage{
\method{autolayer}{recommendation}(
  x,
  color = "green",
  fill = color,
  alpha = 0.2,
  se.fit = T,
  level = 0.95,
  ...
)
}
\arguments{
\item{x}{recommendation}

\item{color}{color to use for the line}

\item{fill}{fill color to use for the ribbon}

\item{alpha}{alpha level for the ribbon}

\item{se.fit}{plot a confidence band around the prediction?}

\item{level}{level for the confidence band}

\item{...}{extra arguments}
}
\description{
Automatically create the required layers for a recommendation object
}
