% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.tdmore}
\alias{autoplot.tdmore}
\title{Plot a tdmore object (typical profile (population) + between subject variability).}
\usage{
\method{autoplot}{tdmore}(x, regimen = NULL, covariates = NULL, ...)
}
\arguments{
\item{x}{the tdmore object}

\item{regimen}{the regimen to be predicted}

\item{covariates}{covariates}

\item{...}{ignored}
}
\value{
a ggplot object with the fitted individual curve, the 95% ci of this curve, the population prediction, the observed data and the residual error around the observed data
}
\description{
Plot a tdmore object (typical profile (population) + between subject variability).
}
