% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.tdmorefit}
\alias{autoplot.tdmorefit}
\title{Plot a tdmorefit object.}
\usage{
\method{autoplot}{tdmorefit}(x, ...)
}
\arguments{
\item{x}{the tdmorefit object}

\item{...}{ignored}
}
\value{
a ggplot object with the fitted individual curve, the 95% ci of this curve, the population prediction, the observed data and the residual error around the observed data
}
\description{
Plot a tdmorefit object.
}
