% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommendation.R
\name{coef.recommendation}
\alias{coef.recommendation}
\title{The obtained parameter values for a recommendation.}
\usage{
\method{coef}{recommendation}(object, ...)
}
\arguments{
\item{object}{a recommendation object}

\item{...}{ignored}
}
\value{
a named numeric vector
}
\description{
The obtained parameter values for a recommendation.
}
