% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{coef.tdmore}
\alias{coef.tdmore}
\title{Get the population typical value parameters. This is a numeric vector of 0's, with the appropriate names.}
\usage{
\method{coef}{tdmore}(object, ...)
}
\arguments{
\item{object}{tdmore object}

\item{...}{ignored}
}
\value{
a numeric vector of 0 values, with the appropriate names
}
\description{
Get the population typical value parameters. This is a numeric vector of 0's, with the appropriate names.
}
