% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{coef.tdmorefit}
\alias{coef.tdmorefit}
\title{The obtained parameter values for the fit}
\usage{
\method{coef}{tdmorefit}(object, ...)
}
\arguments{
\item{object}{A tdmorefit object}

\item{...}{ignored}
}
\value{
Named numeric vector
}
\description{
The obtained parameter values for the fit
}
