% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSpace.R
\name{combinations.searchspace}
\alias{combinations.searchspace}
\title{Add all combinations of the given parameters}
\usage{
combinations.searchspace(searchspace, ...)
}
\arguments{
\item{searchspace}{searchspace object}

\item{...}{All named combinations}
}
\value{
A searchspace with all combinations of the given value added
}
\description{
Add all combinations of the given parameters
}
