% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.autolayer.R
\name{computeMaxTime}
\alias{computeMaxTime}
\title{Compute the maximum interesting time value based on the specified regimen and observed data.}
\usage{
computeMaxTime(regimen, observed = NULL)
}
\arguments{
\item{regimen}{the specified regimen}

\item{observed}{the observed dataframe}
}
\value{
the tmax value, numeric
}
\description{
Compute the maximum interesting time value based on the specified regimen and observed data.
}
\details{
the maximum of all times in observed, and regimen (+one additional Interdose-interval, should this be specified)
}
