% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{computeTmax}
\alias{computeTmax}
\title{Compute the TMax value based on the specified regimen and observed data.}
\usage{
computeTmax(regimen, observed = NULL)
}
\arguments{
\item{regimen}{the specified regimen}

\item{observed}{the observed dataframe}
}
\value{
the tmax value, numeric
}
\description{
Compute the TMax value based on the specified regimen and observed data.
}
