% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{confint.tdmorefit}
\alias{confint.tdmorefit}
\title{Calculates the confidence interval on the obtained parameter values}
\usage{
\method{confint}{tdmorefit}(object, ...)
}
\arguments{
\item{object}{A tdmorefit object}

\item{...}{ignored}
}
\value{
A matrix (or vector) with columns giving
lower and upper confidence limits for each parameter.
These will be labelled as (1-level)/2
and 1 - (1-level)/2 in % (by default 2.5% and 97.5%).
}
\description{
Calculates the confidence interval on the obtained parameter values
}
