% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.qdoptimize.R
\name{convertMCResultToRecommendation}
\alias{convertMCResultToRecommendation}
\title{Convert the Monte-Carlo uniroot results into a recommendation object.}
\usage{
convertMCResultToRecommendation(
  tdmorefit,
  mc,
  result,
  regimen,
  doseRows,
  target,
  level
)
}
\arguments{
\item{mc}{a dataframe with all the monte-carlo samples for parameters}

\item{result}{list with uniroot results}

\item{regimen}{the regimen}

\item{doseRows}{which rows of the regimen to adapt when searching for a new dose, or NULL to take the last one}

\item{target}{the original target}

\item{level}{the confidence interval on the dose}
}
\value{
a recommendation object
}
\description{
Convert the Monte-Carlo uniroot results into a recommendation object.
}
\keyword{internal}
