% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.qdoptimize.R
\name{convertResultToRecommendation}
\alias{convertResultToRecommendation}
\title{Convert the result of the uniroot function into a recommendation object.}
\usage{
convertResultToRecommendation(result, regimen, doseRows, target)
}
\arguments{
\item{result}{the uniroot result}

\item{regimen}{the regimen}

\item{doseRows}{which rows of the regimen to adapt when searching for a new dose, or NULL to take the last one}

\item{target}{the original target}
}
\value{
a recommendation object
}
\description{
Convert the result of the uniroot function into a recommendation object.
}
\keyword{internal}
