% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{covariate}
\alias{covariate}
\title{Create a new covariate object.}
\usage{
covariate(name, label, unit = NULL, min = NULL, max = NULL, choices = NULL)
}
\arguments{
\item{name}{the covariate name}

\item{label}{the covariate label}

\item{unit}{the unit, can be null}

\item{min}{the minimum value, can be null}

\item{max}{the maximum value, can be null}

\item{choices}{a numeric named vector of choices, e.g. c(Fast=0, Slow=1)}
}
\value{
a tdmore_covariate object
}
\description{
Create a new covariate object.
}
\seealso{
Other metadata: 
\code{\link{formulation}()},
\code{\link{getDosingInterval}()},
\code{\link{getMetadataByClass}()},
\code{\link{getMetadataByName}()},
\code{\link{metadata}()},
\code{\link{output}()},
\code{\link{target}()}
}
\concept{metadata}
