% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMixture.R
\name{curry_covariate}
\alias{curry_covariate}
\title{Fill in a specific covariate in a model, transforming it into a more specific model}
\usage{
curry_covariate(m1, curriedCovariates)
}
\description{
Fill in a specific covariate in a model, transforming it into a more specific model
}
