% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dataTibble}
\alias{dataTibble}
\title{Utility function to create a data tibble.}
\usage{
dataTibble(...)
}
\arguments{
\item{...}{each of these arguments will be included in the resulting tibble.}
}
\value{
All arguments passed to the dataTibble function will be present in the resulting tibble as columns.
}
\description{
A data tibble is a tibble with an ID column, and separate columns
with the arguments for `estimate` function calls.
}
