% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModel.R
\name{defaultModel}
\alias{defaultModel}
\title{Get the first available model name in the given directory}
\usage{
defaultModel(dir = system.file(package = pkg, "models"), pkg = "tdmore")
}
\arguments{
\item{dir}{model directory; uses the package `models` directory by default}

\item{pkg}{the package in which to search for a `models` directory}
}
\value{
character string with model filename
}
\description{
Get the first available model name in the given directory
}
