% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSpace.R
\name{discrete.searchspace}
\alias{discrete.searchspace}
\title{Add discrete options to a search space}
\usage{
discrete.searchspace(searchspace, options = NULL)
}
\arguments{
\item{searchspace}{Defined search space}

\item{options}{data.frame with discrete options}
}
\value{
A search space with the discrete options in `options` added
}
\description{
Add discrete options to a search space
}
