% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{dose}
\alias{dose}
\title{Create a new dose object.}
\usage{
dose(unit, name = "DOSE")
}
\arguments{
\item{unit}{the unit}

\item{name}{by default 'DOSE'}
}
\value{
a tdmore_dose object
}
\description{
Create a new dose object.
}
