% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ensurePackagePresent}
\alias{ensurePackagePresent}
\title{This utility function will check if the given package is loaded from a non-standard path.
If so, we are probably running in a sub-process of a test.}
\usage{
ensurePackagePresent(pkgName = "tdmore", quiet = TRUE)
}
\arguments{
\item{pkgName}{the dev package name}

\item{quiet}{FALSE to display log messages}
}
\value{
a directory that can be added to libPaths
}
\description{
If it is available, but in a standard path, then this is probably an outdated version and we need to install the current version.
If it is available, but as a dev_package, then we need to install the current version.
}
\details{
The function always returns a temporary lib directory.
}
\keyword{internal}
