% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorModel.R
\name{errorModel}
\alias{errorModel}
\title{Instantiate a new error model.}
\usage{
errorModel(var, add = 0, prop = 0, exp = 0)
}
\arguments{
\item{var}{the related variable name}

\item{add}{additive residual error, as stdev}

\item{prop}{proportional residual error, as stdev}

\item{exp}{exponential residual error, as stdev. The exponential error cannot be used in conjunction with the additive or proportional error}
}
\description{
Instantiate a new error model.
}
