% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{estimate}
\alias{estimate}
\title{Calculate the Emperical Bayesian Estimated parameters that predict
a given dataset using the model}
\usage{
estimate(tdmore, observed = NULL, regimen, covariates = NULL,
  p = NULL, ...)
}
\arguments{
\item{tdmore}{a tdmore object}

\item{observed}{data frame with at least a TIME column, and all observed data. The observed data will be compared to the model predictions.
If not specified, we estimate the population prediction}

\item{regimen}{data frame describing the treatment regimen.}

\item{covariates}{the model covariates}

\item{p}{optional starting parameter for the MLE minimization}

\item{...}{Extra parameters to pass to nlm}
}
\value{
A tdmorefit object
}
\description{
Calculate the Emperical Bayesian Estimated parameters that predict
a given dataset using the model
}
