% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc.R
\name{estimate.tdmore_mpc}
\alias{estimate.tdmore_mpc}
\title{MPC estimate method.}
\usage{
\method{estimate}{tdmore_mpc}(object, observed = NULL, regimen = NULL,
  covariates = NULL, par = NULL, fix = NULL, method = "L-BFGS-B",
  se.fit = TRUE, lower = NULL, upper = NULL, multistart = F,
  control = list(), data = NULL, ...)
}
\arguments{
\item{object}{a tdmore, tdmore_set or tdmore_mixture object}

\item{observed}{data frame with at least a TIME column, and all observed data. The observed data will be compared to the model predictions.
If not specified, we estimate the population prediction}

\item{regimen}{data frame describing the treatment regimen.}

\item{covariates}{the model covariates}

\item{par}{optional starting parameter for the MLE minimization}

\item{fix}{named vector with the fixed parameters}

\item{method}{the optimisation method, by default, method "L-BFGS-B" is used
Can also be specified as a list, in which case all methods specified will be tried and the one finding the estimate with highest log-likelihood will be returned.}

\item{se.fit}{calculate the variance-covariance matrix for the fit
Putting this to FALSE can reduce computation time.}

\item{lower}{the lower bounds of the parameters, if null, -5 * sqrt(diag(model$omega)) is used}

\item{upper}{the upper bounds of the parameters, if null, +5 * sqrt(diag(model$omega)) is used}

\item{multistart}{if TRUE, perform optimization using many different starting conditions.
A combination of -sd, +sd for each non-IOV parameter is explored.
This avoids returning a local optimum, and strengthens the belief
that the found optimum is a global optimum.

Alternatively, a named vector with the requested perturbations to parameters
can be provided: \code{estimate(multistart=c(eta_Tlag=0.5))}. This will evaluate
the provided starting value for eta_Tlag +- 0.5, without also perturbing the other initial values.

WARNING: This may lead to extremely long computation times. It is recommended to use a control argument to
reduce accuracy and report estimation progress.}

\item{control}{control options for `estimate`
If trace > 0, the estimation function also reports errors when calculating log-likelihood.}

\item{...}{extra parameters to pass to the optim function
A good example is to specify `control=list(trace=1, REPORT=10, factr=1e13)` to improve performance.
This will `trace` the estimation progress
and `report` every 10 iterations. It limits the precision of the `L-BFGS-B` method to 3 significant digits (`factr` is multiplied with the machine precision, `1e12 * 1e-16 = 1e-4`).

Incidentally, 3 significant digits \href{http://holford.fmhs.auckland.ac.nz/research/sigdig}{is the default value for NONMEM}.
Without the above option, `estimate` will calculate with a relative precision of `1e-8`.}
}
\value{
a tdmorefit
}
\description{
MPC estimate method.
}
