% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{expandOmega}
\alias{expandOmega}
\title{Expand OMEGA matrix by adding variance-covariance information regarding all IOV terms.
Note that the column and row names of the returned matrix are strictly identical to the ones returned by getParameterNames().}
\usage{
expandOmega(tdmore, occasions)
}
\arguments{
\item{tdmore}{the tdmore object}

\item{occasions}{how many occasions}
}
\value{
the omega matrix with the duplicated IOV terms at the end
}
\description{
Expand OMEGA matrix by adding variance-covariance information regarding all IOV terms.
Note that the column and row names of the returned matrix are strictly identical to the ones returned by getParameterNames().
}
