% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.qdoptimize.R
\name{findDose}
\alias{findDose}
\title{Find the dose required (quick and dirty, to be optimised)}
\usage{
findDose(tdmorefit, regimen, doseRows = NULL, interval = c(0, 1e+10),
  target, se.fit = FALSE, level = 0.95, mc.maxpts = 100, ...)
}
\arguments{
\item{tdmorefit}{the tdmorefit object}

\item{regimen}{the regimen}

\item{doseRows}{which rows of the regimen to adapt when searching for a new dose, or NULL to take the last one}

\item{interval}{which interval to search a dose in. Defaults to a ridiculously high range}

\item{target}{target value, as a data.frame}

\item{se.fit}{TRUE to provide a confidence interval on the dose prediction, adding columns dose.median, dose.lower and dose.upper}

\item{level}{the confidence interval on the dose, only used if se.fit is true}

\item{mc.maxpts}{maximum number of points to sample in Monte Carlo simulation}

\item{...}{extra arguments passed to stats::uniroot}
}
\value{
a recommendation object
}
\description{
Find the dose required (quick and dirty, to be optimised)
}
