% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{fitted.tdmorefit}
\alias{fitted.tdmorefit}
\title{Get the predicted (fitted) data}
\usage{
\method{fitted}{tdmorefit}(object, ...)
}
\arguments{
\item{object}{A tdmorefit object}

\item{...}{ignored}
}
\value{
A data.frame
}
\description{
Get the predicted (fitted) data
}
