% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regimen.R
\name{flatten}
\alias{flatten}
\title{Flatten a treatment regimen}
\usage{
flatten(regimen)
}
\arguments{
\item{regimen}{data.frame with treatment regimen}
}
\value{
The equivalent treatment regimen with all ADDL objects moved to 0
}
\description{
Flatten a treatment regimen
}
\examples{
flatten(data.frame(TIME=10, AMT=10, ADDL=8, II=24))
}
