% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelLibrary.R
\docType{data}
\name{fluticasone_model}
\alias{fluticasone_model}
\title{Default example model
TODO: Better documentation required}
\format{an nlmixrUI model}
\usage{
fluticasone_model
}
\description{
Default example model
TODO: Better documentation required
}
\keyword{datasets}
