% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{formula.tdmore}
\alias{formula.tdmore}
\title{Get a formula that describes this model}
\usage{
\method{formula}{tdmore}(x, ...)
}
\arguments{
\item{x}{TDMore object}

\item{...}{ignored}
}
\value{
A formula in the form OUTPUT_A + OUTPUT_B ~ TIME + COVARIATES + ETAS
}
\description{
Get a formula that describes this model
}
