% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{formula.tdmorefit}
\alias{formula.tdmorefit}
\title{Get the model used to provide this fit}
\usage{
\method{formula}{tdmorefit}(x, ...)
}
\arguments{
\item{x}{A tdmorefit object}

\item{...}{ignored}
}
\value{
The original tdmore object
}
\description{
Get the model used to provide this fit
}
