% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{fortify.tdmorefit}
\alias{fortify.tdmorefit}
\title{Convert a tdmorefit object to a data.frame using `model.frame`.
The data.frame has a special 'tdmoreArgs' attribute that is used to capture all extra arguments to the original ggplot call.
The attribute contains a list with element 'tdmorefit' and all extra named arguments to the fortify call.}
\usage{
\method{fortify}{tdmorefit}(model, ...)
}
\arguments{
\item{model}{the tdmorefit object}

\item{...}{extra arguments passed to `predict`}
}
\description{
Convert a tdmorefit object to a data.frame using `model.frame`.
The data.frame has a special 'tdmoreArgs' attribute that is used to capture all extra arguments to the original ggplot call.
The attribute contains a list with element 'tdmorefit' and all extra named arguments to the fortify call.
}
