% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{generateMonteCarloMatrix}
\alias{generateMonteCarloMatrix}
\title{Generate the Monte-Carlo matrix.}
\usage{
generateMonteCarloMatrix(tdmorefit, fix, mc.maxpts)
}
\arguments{
\item{tdmorefit}{a tdmorefit object}

\item{fix}{named numeric vector with the fixed parameters}

\item{mc.maxpts}{number of Monte-Carlo samples}
}
\value{
the Monte-Carlo matrix, first column is always the 'sample' column
}
\description{
Generate the Monte-Carlo matrix.
}
