% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{getAes}
\alias{getAes}
\title{This function tries to construct an aes
by guessing the required variables}
\usage{
getAes(x, vars = c("x", "y", "ymin", "ymax"), as.aes = TRUE,
  data = NULL)
}
\arguments{
\item{x}{tdmorefit object}

\item{vars}{the aes vars to guess}

\item{as.aes}{TRUE to return an aes object, FALSE to return a named list}

\item{data}{data.frame to check whether the values actually exist}
}
\description{
If no 'x' is defined, we put a NULL in the aesthetic
It will then get filled in by the predictionLayer
}
