% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{getDosingInterval}
\alias{getDosingInterval}
\title{Get the dosing interval for the given formulation}
\usage{
getDosingInterval(x, model)
}
\arguments{
\item{x}{formulation (character vector)}

\item{model}{a tdmore model with formulation metadata}
}
\description{
Get the dosing interval for the given formulation
}
\examples{
if(FALSE) {
regimen <- data.frame(TIME=0, AMT=5, FORM="CompoundA")
regimen$II <- getDosingInterval(regimen$FORM, model)
}
}
\seealso{
Other metadata: 
\code{\link{covariate}()},
\code{\link{formulation}()},
\code{\link{getMetadataByClass}()},
\code{\link{getMetadataByName}()},
\code{\link{metadata}()},
\code{\link{output}()},
\code{\link{target}()}
}
\concept{metadata}
