% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{getFixedParametersIndexes}
\alias{getFixedParametersIndexes}
\title{Get the indexes of the fixed parameters.}
\usage{
getFixedParametersIndexes(par, fix)
}
\arguments{
\item{par}{named vector with the parameters in the right order}

\item{fix}{named vector with the fixed parameters (user input)}
}
\value{
the fixed parameters indexes
}
\description{
Get the indexes of the fixed parameters.
}
