% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{getLikelihoodFun}
\alias{getLikelihoodFun}
\title{Get the right likelihood function for the specified type.}
\usage{
getLikelihoodFun(type)
}
\arguments{
\item{type}{likelihood type, ll, pop or pred}
}
\value{
the correct tdmore function
}
\description{
Get the right likelihood function for the specified type.
}
