% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regimen.R
\name{getMaxOccasion}
\alias{getMaxOccasion}
\title{Get the highest occasion number in a regimen.
If the regimen does not contain an 'OCC' column, the number of rows in the regimen is returned (each regimen a separate occasion)}
\usage{
getMaxOccasion(regimen)
}
\arguments{
\item{regimen}{the specified regimen}
}
\value{
the highest occasion number
}
\description{
Get the highest occasion number in a regimen.
If the regimen does not contain an 'OCC' column, the number of rows in the regimen is returned (each regimen a separate occasion)
}
