% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{getMetadataByClass}
\alias{getMetadataByClass}
\title{Get metadata by class.}
\usage{
getMetadataByClass(tdmore, metaClass)
}
\arguments{
\item{tdmore}{the tdmore model}

\item{metaClass}{the class of the metadata}
}
\value{
the metadata items that matches the given class
}
\description{
Get metadata by class.
}
