% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{getMetadataByName}
\alias{getMetadataByName}
\title{Get metadata (output, formulation or covariate) by name.}
\usage{
getMetadataByName(tdmore, metaName)
}
\arguments{
\item{tdmore}{the tdmore model}

\item{metaName}{the name of the metadata}
}
\value{
the first metadata item that matches the given name
}
\description{
Get metadata (output, formulation or covariate) by name.
}
