% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModel.R
\name{getModel}
\alias{getModel}
\title{This searches for the given model name in a directory}
\usage{
getModel(
  name = defaultModel(),
  dir = system.file(package = pkg, "models"),
  pkg = "tdmore"
)
}
\arguments{
\item{name}{model name (filename without `.r`)}

\item{dir}{model directory; uses the package `models` directory by default}

\item{pkg}{the package in which to search for a `models` directory}
}
\value{
a tdmore model object
}
\description{
This searches for the given model name in a directory
}
