% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{getObservedVariables}
\alias{getObservedVariables}
\title{Get the observed variables from the tdmore model.}
\usage{
getObservedVariables(model)
}
\arguments{
\item{model}{tdmore model}
}
\value{
a character vector of observed variables
}
\description{
Get the observed variables from the tdmore model.
}
