% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regimen.R
\name{getOccasionTimes}
\alias{getOccasionTimes}
\title{Get the times related to each occasion in a regimen.}
\usage{
getOccasionTimes(regimen)
}
\arguments{
\item{regimen}{the specified regimen}
}
\value{
a numeric vector with the times
}
\description{
Get the times related to each occasion in a regimen.
}
