% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{getParameterNames}
\alias{getParameterNames}
\title{Get the parameter names. If IOV is defined in the model,
IOV terms will be added at the end of the character array and repeated as many times as there are occasions in the regimen.
E.g. 2 occasions: "ECL", "EKA", "ECL_IOV", "EKA_IOV", "ECL_IOV", "EKA_IOV"}
\usage{
getParameterNames(tdmore, regimen)
}
\arguments{
\item{tdmore}{the tdmore object}

\item{regimen}{the regimen}
}
\value{
an ordered character array with the parameters names
}
\description{
Get the parameter names. If IOV is defined in the model,
IOV terms will be added at the end of the character array and repeated as many times as there are occasions in the regimen.
E.g. 2 occasions: "ECL", "EKA", "ECL_IOV", "EKA_IOV", "ECL_IOV", "EKA_IOV"
}
