% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{getPredictOutputNames}
\alias{getPredictOutputNames}
\title{Return the output names of the fitted MC matrix.}
\usage{
getPredictOutputNames(newdata, columnNames, pNames)
}
\arguments{
\item{newdata}{the newdata data frame}

\item{pNames}{parameters names}

\item{colnames}{current column names}
}
\value{
a vector with all the output names
}
\description{
Return the output names of the fitted MC matrix.
}
