% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggplot.tdmorefit}
\alias{ggplot.tdmorefit}
\title{Create a ggplot object from a `tdmorefit` object.
The plot can be used as normal, but also supports adding specific tdmore layers, such as `ipred`.}
\usage{
\method{ggplot}{tdmorefit}(data = NULL, mapping = aes(),
  newdata = NULL, regimen = NULL, parameters = NULL,
  covariates = NULL, ..., environment = parent.frame())
}
\arguments{
\item{data}{A tdmorefit object}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{newdata}{A data.frame with new data and the columns to predict,
or a numeric vector to specify times, and predict all model output
or NULL to interpolate between 0 and the maximum known times}

\item{regimen}{Treatment regimen}

\item{parameters}{Set parameters. If missing, or if only part of the parameters are specified,
the other parameters are taken from the tdmorefit object}

\item{covariates}{the model covariates, named vector, or data.frame with column 'TIME', and at least TIME 0}

\item{...}{Other arguments passed on to methods. Not currently used.}

\item{environment}{DEPRECATED. Used prior to tidy evaluation.}
}
\description{
Create a ggplot object from a `tdmorefit` object.
The plot can be used as normal, but also supports adding specific tdmore layers, such as `ipred`.
}
