% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggplot_add.GGTdmoreLayer}
\alias{ggplot_add.GGTdmoreLayer}
\title{This method extends ggplot2 to allow inheritance of tdmore arguments
from the original `ggplot()` call into the specific tdmore-related layer.}
\usage{
\method{ggplot_add}{GGTdmoreLayer}(object, plot, object_name)
}
\arguments{
\item{object}{An object to add to the plot}

\item{plot}{The ggplot object to add \code{object} to}

\item{object_name}{The name of the object to add}
}
\description{
This method extends ggplot2 to allow inheritance of tdmore arguments
from the original `ggplot()` call into the specific tdmore-related layer.
}
\keyword{internal}
