% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regimen.R
\name{hasOccasion}
\alias{hasOccasion}
\title{Tell if a regimen has an occasion column.}
\usage{
hasOccasion(regimen)
}
\arguments{
\item{regimen}{the specified regimen}
}
\value{
TRUE if the regimen has a 'OCC' column, false otherwise
}
\description{
Tell if a regimen has an occasion column.
}
