% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc.R
\name{is.mpc}
\alias{is.mpc}
\title{Test if an object is a `tdmore_mpc` model}
\usage{
is.mpc(x)
}
\arguments{
\item{x}{object to test}
}
\value{
TRUE if x inherits from `tdmore_mpc`
}
\description{
Test if an object is a `tdmore_mpc` model
}
