% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{is.tdmorefit}
\alias{is.tdmorefit}
\title{Check if the given argument is of class `tdmorefit`}
\usage{
is.tdmorefit(a)
}
\arguments{
\item{x}{An object to test}
}
\description{
Check if the given argument is of class `tdmorefit`
}
\keyword{internal}
