% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModel.R
\name{listModels}
\alias{listModels}
\title{This lists the models in a directory}
\usage{
listModels(dir = system.file(package = pkg, "models"), pkg = "tdmore")
}
\arguments{
\item{dir}{model directory; uses the package `models` directory by default}

\item{pkg}{the package in which to search for a `models` directory}
}
\value{
a character vector of available model names
}
\description{
This lists the models in a directory
}
