% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{ll}
\alias{ll}
\title{Calculate the log likelihood as the sum of the population likelihood and the prediction likelihood.}
\usage{
ll(par, tdmore, observed, regimen, covariates)
}
\arguments{
\item{par}{the current estimate of the parameters}

\item{tdmore}{the tdmore object}

\item{observed}{data frame with at least a TIME column, and all observed data. The observed data will be compared to the model predictions.
If not specified, we estimate the population prediction}

\item{regimen}{data frame describing the treatment regimen.}

\item{covariates}{the model covariates}
}
\value{
the log likelihood
}
\description{
Calculate the log likelihood as the sum of the population likelihood and the prediction likelihood.
}
