% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{logLik.tdmorefit}
\alias{logLik.tdmorefit}
\title{Get the log-likelihood of the predicted values.}
\usage{
\method{logLik}{tdmorefit}(object, type = c("ll", "pop", "pred"), ...)
}
\arguments{
\item{object}{A tdmorefit object}

\item{type}{log-lokelihood function type, 3 possible values: 'pop', 'pred' or 'll' (= pop + pred)}

\item{...}{ignored}
}
\value{
A numeric value
}
\description{
Get the log-likelihood of the predicted values.
}
