% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{meltPredictions}
\alias{meltPredictions}
\title{Melt prediction results.}
\usage{
meltPredictions(x, se = FALSE)
}
\arguments{
\item{x}{dataframe returned by the predict function}

\item{se}{if FALSE, melts the data.frame as expected
if TRUE, assumes every output has 4 columns in the data.frame: XXX1, XXX1.median, XXX1.lower, XXX1.upper
And melts these into a variable (with values XXX1, XXX2, ...) and 4 columns: value, value.median, value.lower and value.upper}
}
\value{
the melted dataframe
}
\description{
Melt prediction results.
}
\keyword{internal}
