% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc.R
\name{mergeCovariates}
\alias{mergeCovariates}
\title{Merge covariates.}
\usage{
mergeCovariates(ebeCovariates, covariates)
}
\arguments{
\item{ebeCovariates}{EBE covariates}

\item{covariates}{'true' time-varying covariates}
}
\value{
an updated covariates dataframe
}
\description{
Merge covariates.
}
