% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-models.R
\docType{data}
\name{meropenem_model}
\alias{meropenem_model}
\title{Default example model
TODO: Better documentation required}
\format{an nlmixrUI model}
\usage{
meropenem_model
}
\description{
Default example model
TODO: Better documentation required
}
\keyword{datasets}
