% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata.default}
\alias{metadata.default}
\title{Default function, catching unsupported types of metadata}
\usage{
\method{metadata}{default}(x, ...)
}
\arguments{
\item{x}{anything}

\item{...}{ignored}
}
\description{
Default function, catching unsupported types of metadata
}
