% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{metadata is a generic function to append metadata to a TDMore model.}
\usage{
metadata(...)
}
\arguments{
\item{...}{extra arguments}
}
\value{
an object of class tdmore
}
\description{
metadata is a generic function to append metadata to a TDMore model.
}
