% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata.tdmore}
\alias{metadata.tdmore}
\title{Append metadata to a TDMore model.}
\usage{
\method{metadata}{tdmore}(tdmore, ...)
}
\arguments{
\item{tdmore}{a tdmore_output object}

\item{...}{metadata}
}
\value{
a tdmore model
}
\description{
Append metadata to a TDMore model.
}
