% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata.tdmore_covariate}
\alias{metadata.tdmore_covariate}
\title{Build a covariate metadata.}
\usage{
\method{metadata}{tdmore_covariate}(x, ...)
}
\arguments{
\item{x}{a tdmore_covariate object}

\item{...}{ignored}
}
\value{
a metadata structure
}
\description{
Build a covariate metadata.
}
