% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata.tdmore_dose}
\alias{metadata.tdmore_dose}
\title{Build a dose metadata.}
\usage{
\method{metadata}{tdmore_dose}(x, ...)
}
\arguments{
\item{x}{a tdmore_dose object}

\item{...}{ignored}
}
\value{
a metadata structure
}
\description{
Build a dose metadata.
}
