% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata.tdmore_formulation}
\alias{metadata.tdmore_formulation}
\title{Build a formulation metadata.}
\usage{
\method{metadata}{tdmore_formulation}(x, ...)
}
\arguments{
\item{x}{a tdmore_formulation object}

\item{...}{ignored}
}
\value{
a metadata structure
}
\description{
Build a formulation metadata.
}
