% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiMetadata.R
\name{metadata.tdmore_observed_variables}
\alias{metadata.tdmore_observed_variables}
\title{Build an 'observed variables' metadata.}
\usage{
\method{metadata}{tdmore_observed_variables}(x, ...)
}
\arguments{
\item{x}{a tdmore_observed_variables object}

\item{...}{ignored}
}
\value{
a metadata structure
}
\description{
Build an 'observed variables' metadata.
}
