% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata.tdmore_output}
\alias{metadata.tdmore_output}
\title{Build an output metadata.}
\usage{
\method{metadata}{tdmore_output}(x, ...)
}
\arguments{
\item{x}{a tdmore_output object}

\item{...}{ignored}
}
\value{
a metadata structure
}
\description{
Build an output metadata.
}
