% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata.tdmore_target}
\alias{metadata.tdmore_target}
\title{Build a target metadata.}
\usage{
\method{metadata}{tdmore_target}(x, ...)
}
\arguments{
\item{x}{a tdmore_target object}

\item{...}{ignored}
}
\value{
a metadata structure
}
\description{
Build a target metadata.
}
