% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.frame.tdmore}
\alias{model.frame.tdmore}
\title{Get the specified observed values, with the upper and lower bounds provided by the residual error model}
\usage{
\method{model.frame}{tdmore}(formula, observed, se = FALSE,
  level = 0.95, ...)
}
\arguments{
\item{formula}{a TDMore object}

\item{observed}{data.frame with at least a TIME column}

\item{se}{TRUE to generate an additional xx.upper and xx.lower column for every xx column in the observed dataset}

\item{level}{numeric value to specify the confidence interval}

\item{...}{ignored}
}
\value{
a data.frame similar to observed, but with additional columns xx.lower and xx.upper (in case of se=TRUE)
}
\description{
Get the specified observed values, with the upper and lower bounds provided by the residual error model
}
