% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{model.frame.tdmorefit}
\alias{model.frame.tdmorefit}
\title{Get the observed values used to provide this model fit.}
\usage{
\method{model.frame}{tdmorefit}(formula, se = FALSE, level = 0.95, ...)
}
\arguments{
\item{formula}{A tdmorefit object}

\item{se}{if TRUE, add a column xx.upper and xx.lower with the lower and upper bounds of confidence, based on the residual error model}

\item{level}{Confidence interval}

\item{...}{ignored}
}
\value{
A data.frame, similar to the original
}
\description{
Get the observed values used to provide this model fit.
}
