% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraic.R
\name{model_predict.algebraic}
\alias{model_predict.algebraic}
\title{Predict for algebraic models}
\usage{
\method{model_predict}{algebraic}(
  model,
  times,
  regimen = NULL,
  parameters = numeric(),
  covariates = NULL,
  iov = NULL,
  extraArguments = list(),
  cache = NULL
)
}
\arguments{
\item{model}{the algebraic model}

\item{times}{numeric vector at which to provide prediction values}

\item{regimen}{dataframe with column 'TIME' and adhering to standard NONMEM specifications otherwise (columns AMT, RATE, CMT)}

\item{parameters}{a named numeric vector}

\item{covariates}{named numeric vector, or data.frame with column 'TIME', and at least TIME 0}

\item{iov}{character array with the IOV terms, NULL if no IOV. These values should then be present in the parameters vector, and be repeated for
max(regimen$OCC).}
}
\value{
A data.frame containing all output of the prediction function of the algebraic model
}
\description{
Predict for algebraic models
}
\keyword{internal}
