% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{model_predict.RxODE}
\alias{model_predict.RxODE}
\title{Predict for RxODE}
\usage{
\method{model_predict}{RxODE}(model, times, regimen = data.frame(TIME =
  numeric()), parameters = numeric(), covariates = NULL,
  extraArguments = list())
}
\arguments{
\item{model}{The model itself.}

\item{regimen}{dataframe with column 'TIME' and adhering to standard NONMEM specifications otherwise (columns AMT, RATE, CMT).}

\item{parameters}{named numeric vector with the parameters}

\item{covariates}{named vector, or data.frame with column 'TIME', and at least TIME 0}

\item{extraArguments}{extra arguments to use}

\item{newdata}{A dataframe with at least a column 'TIME' and other values that should be observed,
or a numeric TIME vector to produce all values that can be observed by this model}
}
\value{
A data.frame similar to the observed data frame, but with predicted values.
}
\description{
Predict for RxODE
}
\keyword{internal}
