% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deSolve.R
\name{model_predict.tdmore_deSolve}
\alias{model_predict.tdmore_deSolve}
\title{Predict for tdmore_deSolve}
\usage{
\method{model_predict}{tdmore_deSolve}(model, newdata,
  regimen = data.frame(TIME = c()), parameters = c(),
  covariates = NULL, extraArguments = list())
}
\arguments{
\item{model}{The model function}

\item{newdata}{data frame with at least a TIME column, and all observed data. The observed data will be compared to the model predictions.
If not specified, we estimate the population prediction}

\item{regimen}{dataframe with column 'TIME' and adhering to standard NONMEM specifications otherwise (columns AMT, RATE, CMT).}

\item{parameters}{either a dataframe with column 'TIME' and a column for each covariate and parameter, or a named numeric vector}

\item{covariates}{NOT IMPLEMENTED}

\item{extraArguments}{extra arguments to use}
}
\value{
A data.frame similar to the observed data frame, but with predicted values.
}
\description{
Predict for tdmore_deSolve
}
