% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_prepare}
\alias{model_prepare}
\title{Prepare a cache object}
\usage{
model_prepare(model, times, regimen, parameters, covariates, iov,
  extraArguments)
}
\arguments{
\item{model}{The model itself.}

\item{times}{The times at which to generate predictions. May be empty, but never NULL.}

\item{regimen}{dataframe with column 'TIME' and adhering to standard NONMEM specifications otherwise (columns AMT, RATE, CMT)}

\item{parameters}{named vector}

\item{covariates}{named vector, or data.frame with column 'TIME', and at least TIME 0}

\item{iov}{IOV terms}

\item{extraArguments}{named list with extra arguments to use for call}
}
\value{
A list of cached values
}
\description{
Prepare a cache object
}
\keyword{internal}
