% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{model_prepare.RxODE}
\alias{model_prepare.RxODE}
\title{Prepare a cache object for RxODE simulation}
\usage{
\method{model_prepare}{RxODE}(model, times, regimen = data.frame(TIME =
  numeric()), parameters = numeric(), covariates = NULL, iov = NULL,
  extraArguments = list())
}
\arguments{
\item{model}{The model itself.}

\item{regimen}{dataframe with column 'TIME' and adhering to standard NONMEM specifications otherwise (columns AMT, RATE, CMT).}

\item{parameters}{named numeric vector with the parameters}

\item{covariates}{named vector, or data.frame with column 'TIME', and at least TIME 0}

\item{iov}{character array with the parameters that should change for every different OCC, can be NULL}

\item{extraArguments}{extra arguments to use}

\item{newdata}{A dataframe with at least a column 'TIME' and other values that should be observed,
or a numeric TIME vector to produce all values that can be observed by this model}
}
\value{
A list, to be used as a cache object for model_predict
}
\description{
Prepare a cache object for RxODE simulation
}
\keyword{internal}
