% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc.R
\name{mpc}
\alias{mpc}
\title{MPC is a generic function to make a tdmore model compatible with MPC.}
\usage{
mpc(x, theta, suffix, ...)
}
\arguments{
\item{x}{a model object (type not defined)}

\item{theta}{named vector with the initial values}

\item{suffix}{a suffix to add to the theta names}

\item{...}{extra arguments}
}
\value{
an object of class tdmore_mpc
}
\description{
MPC is a generic function to make a tdmore model compatible with MPC.
}
