% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc.R
\name{mpc.tdmore}
\alias{mpc.tdmore}
\title{Build a 'tdmore_mpc' object based on a 'tdmore' object.}
\usage{
\method{mpc}{tdmore}(x, parameters, ...)
}
\arguments{
\item{x}{a model object (type not defined)}

\item{parameters}{select which parameters should be mpc parameters. MPC parameters are cumulatively summed across occasions.}

\item{...}{extra arguments}
}
\value{
an object of class tdmore_mpc
}
\description{
Build a 'tdmore_mpc' object based on a 'tdmore' object.
}
