% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiMetadata.R
\name{observed_variables}
\alias{observed_variables}
\title{Create an 'observed variables' object.
These variables will be predicted and shown in the panel below the timeline.
TODO: discuss if we keep this info as part of the UI metadata of the model.
Should we need a config tab to configure the UI metadata?
Do we want to persist this information in DB? Per model? Per patient? etc.}
\usage{
observed_variables(variables)
}
\arguments{
\item{variables}{variables to be observed}
}
\value{
a tdmore_observed_variables object
}
\description{
Create an 'observed variables' object.
These variables will be predicted and shown in the panel below the timeline.
TODO: discuss if we keep this info as part of the UI metadata of the model.
Should we need a config tab to configure the UI metadata?
Do we want to persist this information in DB? Per model? Per patient? etc.
}
