% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{output}
\alias{output}
\title{Create a new output object.}
\usage{
output(name, label, unit = NULL)
}
\arguments{
\item{name}{the output name}

\item{label}{the output label}

\item{unit}{the unit, can be null}
}
\value{
a tdmore_output object
}
\description{
Create a new output object.
}
