% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelLibrary.R
\docType{data}
\name{pheno}
\alias{pheno}
\title{Phenobarbitol dataset, as included in the NONMEM distribution.}
\format{
A data frame with 744 rows
\describe{
  \item{ID}{Subject ID}
  \item{TIME}{Time of observation, in days after first dose}
  \item{AMT}{phenobarbitol dose, in mg}
  \item{WT}{Weight in kg}
  \item{APGR}{APGAR score, a measure for sedation}
  \item{DV}{Concentration, in mg/L}
  \item{EVID}{1 for administration, 0 for observation}
  \item{MDV}{1 for administration, 0 for observation}
}
}
\usage{
pheno
}
\description{
This is comprised of 59 subjects with sparse sampled concentrations.
There are 155 measured concentrations and 589 administrations.
}
\keyword{datasets}
