% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelLibrary.R
\docType{data}
\name{pheno_nlmixr}
\alias{pheno_nlmixr}
\title{NlmixrUI object describing Phenobarbitol 1-compartment model with IV administration
and allometric scaling on weight}
\format{
A fitted Nlmixr model
}
\usage{
pheno_nlmixr
}
\description{
The model was estimated by nlmixr SAEM
}
\keyword{datasets}
