% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraicLibrary.R
\name{pk}
\alias{pk}
\alias{pk1cptiv}
\alias{pk1cptinfusion}
\alias{pk1cptoral}
\alias{pk2cptiv}
\alias{pk2cptinfusion}
\alias{pk2cptoral}
\alias{pk3cptiv}
\alias{pk3cptinfusion}
\alias{pk3cptoral}
\title{Executes the requested PK model, and
fetches the arguments from the caller's environment.}
\usage{
pk1cptiv(fName = as.character(sys.call())[1], env = parent.frame())

pk1cptinfusion(fName = as.character(sys.call())[1],
  env = parent.frame())

pk1cptoral(fName = as.character(sys.call())[1], env = parent.frame())

pk2cptiv(fName = as.character(sys.call())[1], env = parent.frame())

pk2cptinfusion(fName = as.character(sys.call())[1],
  env = parent.frame())

pk2cptoral(fName = as.character(sys.call())[1], env = parent.frame())

pk3cptiv(fName = as.character(sys.call())[1], env = parent.frame())

pk3cptinfusion(fName = as.character(sys.call())[1],
  env = parent.frame())

pk3cptoral(fName = as.character(sys.call())[1], env = parent.frame())
}
\arguments{
\item{fName}{function name to delegate this call too
filled in automatically}

\item{env}{environment in which to search the function arguments
filled in automatically}
}
\description{
Executes the requested PK model, and
fetches the arguments from the caller's environment.
}
