% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraicLibrary.R
\name{pk1cptoralbolusVCL}
\alias{pk1cptoralbolusVCL}
\title{A tdmore model to predict concentration in a 1cpt PK model with bolus administration}
\usage{
pk1cptoralbolusVCL(TVKA = 0.5, TVV = 10, TVCL = 5, OKA = 0.7,
  OV = 0.2, OCL = 0.3, res_var)
}
\arguments{
\item{TVKA}{Typical value for absorption constant}

\item{TVV}{Typical value for volume}

\item{TVCL}{Typical value for clearance}

\item{OKA}{Typical value for clearance}

\item{OV}{Omega for volume}

\item{OCL}{Omega for clearance}

\item{res_var}{Residual error model}
}
\value{
a tdmore model
}
\description{
A tdmore model to predict concentration in a 1cpt PK model with bolus administration
}
\examples{
m1 <- pk1cptoralbolusVCL(
  TVKA=0.5, TVV=10, TVCL=5,
  OKA=0.70, OV=0.20, OCL=0.30,
  errorModel(prop=0.1)
)
summary(m1)
}
