% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraicLibrary.R
\name{pk2cptivinfusion}
\alias{pk2cptivinfusion}
\title{This base function can be used to develop an algebraic model
for a 2cpt pk model with oral absorption.}
\usage{
pk2cptivinfusion()
}
\description{
The caller should have access to the following values:
t, TIME, AMT
as well as model parameters
CL, V1, Q, V2
}
\details{
The following alternative parametrization is also permitted
K12, K21, K

Optionally, TLAG can be defined.
}
\examples{
m1 <- algebraic(function(t, TIME, AMT, EV, ECL) {
    KA <- 0.3
    CL <- 3 * exp(ECL)
    V1 <- 10 * exp(EV)
    Q <- 3
    V2 <- 800
    DUR <- 3
    pk2cptoralinfusion()
})
}
