% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraicLibrary.R
\name{pk_}
\alias{pk_}
\alias{pk1cptiv_}
\alias{pk1cptinfusion_}
\alias{pk1cptoral_}
\alias{pk2cptiv_}
\alias{pk2cptinfusion_}
\alias{pk2cptoral_}
\alias{pk3cptiv_}
\alias{pk3cptinfusion_}
\alias{pk3cptoral_}
\title{Algebraic equations for PK models}
\usage{
pk1cptiv_(t, TIME, AMT, V, K, SS = 0, II = Inf)

pk1cptinfusion_(t, TIME, AMT, RATE, V, K, SS = 0, II = Inf)

pk1cptoral_(t, TIME, AMT, V, K, KA, SS = 0, II = Inf)

pk2cptiv_(t, TIME, AMT, V, K, K12, K21, SS = 0, II = Inf)

pk2cptinfusion_(t, TIME, AMT, RATE, V, K, K12, K21, SS = 0, II = Inf)

pk2cptoral_(t, TIME, AMT, V, K, KA, K12, K21, SS = 0, II = Inf)

pk3cptiv_(t, TIME, AMT, V, K, K12, K21, K13, K31, SS = 0, II = Inf)

pk3cptinfusion_(t, TIME, AMT, RATE, V, K, K12, K21, K13, K31, SS = 0,
  II = Inf)

pk3cptoral_(t, TIME, AMT, V, K, KA, K12, K21, K13, K31, SS = 0,
  II = Inf)
}
\arguments{
\item{t}{sampling times}

\item{V}{volume of distribution}

\item{K}{elimination constant}

\item{tD}{dosing time}

\item{D}{dosing amount}
}
\value{
vector of same size as `t`, with the concentrations
}
\description{
Algebraic equations for PK models
}
