% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraicLibrary.R
\name{pkmodel}
\alias{pkmodel}
\title{Execute the appropriate PK model, based on the same rules as used by Monolix}
\usage{
pkmodel(...)
}
\arguments{
\item{...}{the parameters for the PK model}
}
\description{
Execute the appropriate PK model, based on the same rules as used by Monolix
}
\details{
See http://monolix.lixoft.com/data-and-models/pkmodel1/ for more information
}
