% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tdmore}
\alias{plot.tdmore}
\title{Plot a tdmore object (typical profile of the population and between subject variability).}
\usage{
\method{plot}{tdmore}(x, regimen, covariates = NULL, newdata = NULL,
  bsv = TRUE, mc.subjects = 100, ...)
}
\arguments{
\item{x}{the tdmore object}

\item{regimen}{the regimen to be predicted}

\item{covariates}{covariates}

\item{newdata}{a data.frame with at least TIME and any other columns to plot, NULL to plot all columns from the original observed data between time 0 and max(observationTime) or a numeric vector of times}

\item{bsv}{add between subject variability, enabled by default}

\item{mc.subjects}{number of subjects in the monte carlo simulation}

\item{...}{ignored}
}
\value{
a ggplot object with the the population prediction
}
\description{
Plot a tdmore object (typical profile of the population and between subject variability).
}
