% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tdmorefit}
\alias{plot.tdmorefit}
\title{Plot a tdmorefit object.}
\usage{
\method{plot}{tdmorefit}(x, newdata = NULL, regimen = NULL,
  se.fit = TRUE, population = TRUE, fit = TRUE, mc.maxpts = 100,
  .progress = "none", ...)
}
\arguments{
\item{x}{the tdmorefit object}

\item{newdata}{a data.frame with at least TIME and any other columns to plot, NULL to plot all columns from the original observed data between time 0 and max(observationTime) or a numeric vector of times}

\item{regimen}{Treatment regimen}

\item{se.fit}{add a curve for the confidence interval around the fit}

\item{population}{should I plot the population prediction (in blue) ?}

\item{fit}{should I plot the fit (in red) ?}

\item{mc.maxpts}{maximum number of points to use for the monte carlo fit}

\item{.progress}{either "none" or "text" to see calculation progress of monte carlo simulations}

\item{...}{ignored}
}
\value{
a ggplot object with the fitted individual curve, the 95% CI of this curve, the population prediction (with between-subject variability) and the observed data
}
\description{
Plot a tdmorefit object.
}
