% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tdmorefit_mixture}
\alias{plot.tdmorefit_mixture}
\title{Plot a tdmorefit mixture object.}
\usage{
\method{plot}{tdmorefit_mixture}(x, newdata = NULL, regimen = NULL,
  se.fit = TRUE, population = TRUE, fit = TRUE, mc.maxpts = 100,
  .progress = "none", ...)
}
\arguments{
\item{x}{the tdmorefit mixture object}

\item{newdata}{A data.frame with new data and the columns to predict,
or a numeric vector to specify times, and predict all model output
or NULL to interpolate between 0 and the maximum known times}

\item{regimen}{Treatment regimen}

\item{se.fit}{TRUE to provide a confidence interval on the prediction, adding columns xxx.median, xxx.upper and xxx.lower}

\item{mc.maxpts}{Maximum number of points to sample in Monte Carlo simulation}

\item{.progress}{see plyr::ddply}

\item{...}{ignored}
}
\description{
Plot a tdmorefit mixture object.
}
