% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{plot.tdmoreprofile}
\alias{plot.tdmoreprofile}
\title{Plot a tdmoreprofile object.}
\usage{
\method{plot}{tdmoreprofile}(x, parameters = NULL, contour = T,
  raster = T, ...)
}
\arguments{
\item{x}{the tdmoreprofile object}

\item{parameters}{the parameters to plot}

\item{contour}{use the 'geom_contour' method to draw the profile, logical value}

\item{raster}{use the 'raster method' to draw the profile, logical value}

\item{...}{unused arguments}
}
\description{
Plot a tdmoreprofile object.
}
