% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{plotpossibilities}
\alias{plotpossibilities}
\title{Plot the fitted prediction under all discrete regimen possibilities}
\usage{
plotpossibilities(tdmorefit, regimen, searchspace, evaluators)
}
\arguments{
\item{tdmorefit}{A tdmore fit}

\item{regimen}{Candidate regimen}

\item{searchspace}{a list with searchSpaces for every row in the regimen table, or a single searchspace
In case this is a single searchSpace, it will be attributed to
every row in the regimen table after the last observation.
Any dose rows where ADDL is spanning over the last observation will
not be attributed to the searchSpace.}

\item{evaluators}{named list of evaluation functions(tdmorefit, regimen)}
}
\value{
list with source data for plots, as well as the plots in question
}
\description{
Plot the fitted prediction under all discrete regimen possibilities
}
