% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{pop_ll}
\alias{pop_ll}
\title{Calculate the population log likelihood.}
\usage{
pop_ll(par, omega, fix, tdmore, observed, regimen, covariates,
  isChol = FALSE)
}
\arguments{
\item{par}{the current estimate of the parameters}

\item{omega}{the omega matrix}

\item{fix}{named vector with the fixed parameters, not used in this function}

\item{tdmore}{the tdmore object}

\item{observed}{the observed data, not used in the population log likelihood}

\item{regimen}{data frame describing the treatment regimen}

\item{covariates}{the model covariates, not used in the population log likelihood}

\item{isChol}{was omega specified as the Cholesky decomposition?}
}
\value{
the population log likelihood
}
\description{
Calculate the population log likelihood.
}
