% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSpace.R
\name{possibilitygrid}
\alias{possibilitygrid}
\title{Calculate all possible regimen combinations, as rows in a data.frame}
\usage{
possibilitygrid(searchspace)
}
\arguments{
\item{searchspace}{searchspace object}
}
\value{
A data.frame with e.g. columns 1.Foo, 1.Bar 2.Foo,
with the rows containing all discrete combinations for the Foo value in the first row,
the Bar value in the first row and the Foo value in the second row.
}
\description{
Calculate all possible regimen combinations, as rows in a data.frame
}
