% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proseval.R
\name{posthoc}
\alias{posthoc}
\title{This function performs a posthoc estimate for multiple individuals at the same time.
All parameters may have an additional ID column to identify every individual.}
\usage{
posthoc(object, observed = NULL, regimen = NULL, covariates = NULL,
  par = NULL, fix = NULL, ..., .prediction = "ipred")
}
\arguments{
\item{object}{a tdmore, tdmore_set or tdmore_mixture object}

\item{observed}{data frame with at least a TIME column, and all observed data. The observed data will be compared to the model predictions.
If not specified, we estimate the population prediction}

\item{regimen}{data frame describing the treatment regimen.}

\item{covariates}{the model covariates}

\item{par}{optional starting parameter for the MLE minimization}

\item{fix}{named vector with the fixed parameters}

\item{...}{passed to \link{estimate} function}

\item{.prediction}{Either a string or NULL. If a string, the output will contain a column
with that name, storing the predicted values for that specific fit.}
}
\value{
a tibble with columns ID,
and column `fit` with the tdmorefit object
}
\description{
This function performs a posthoc estimate for multiple individuals at the same time.
All parameters may have an additional ID column to identify every individual.
}
