% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{predict.tdmore}
\alias{predict.tdmore}
\title{Predict from a tdmore model}
\usage{
\method{predict}{tdmore}(object, newdata, regimen, parameters = NULL,
  covariates = NULL, se = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{TDMore object}

\item{newdata}{Data.frame of new data with at least a TIME column, and blank columns with all desired predictions
or a numeric vector to predict all possible values from the model}

\item{regimen}{Treatment regimen}

\item{parameters}{The parameter values to use, missing values are taken from the population}

\item{covariates}{the model covariates, named vector, or data.frame with column 'TIME', and at least TIME 0}

\item{se}{Whether to add residual error}

\item{level}{How much residual error to add}

\item{...}{ignored}
}
\value{
A data.frame with all observed values at the given time points
}
\description{
Predict from a tdmore model
}
\seealso{
tdmore::model_predict
}
