% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc.R
\name{predict.tdmore_mpc}
\alias{predict.tdmore_mpc}
\title{MPC predict method. This simply generates a cumulative sum for all}
\usage{
\method{predict}{tdmore_mpc}(
  object,
  newdata,
  regimen = NULL,
  parameters = NULL,
  covariates = NULL,
  se = FALSE,
  level = 0.95,
  ...
)
}
\value{
a data.frame with all observed values at the given time points
}
\description{
MPC predict method. This simply generates a cumulative sum for all
}
