% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMixture.R
\name{predict.tdmoreCurried}
\alias{predict.tdmoreCurried}
\title{Predict a tdmoreCurried object.}
\usage{
\method{predict}{tdmoreCurried}(object, ...)
}
\arguments{
\item{object}{a tdmoreCurried object}

\item{...}{all the extra arguments to pass to the next predict method}
}
\value{
a dataframe with the predictions
}
\description{
Predict a tdmoreCurried object.
}
