% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{predictionLayer}
\alias{predictionLayer}
\title{This creates a special type of ggplot2 Layer object.
The layer searches for a tdmorefit object in either the plot, or its own arguments.
It then calls the `predict()` function and provides this data.frame instead of the original data}
\usage{
predictionLayer(
  mapping = NULL,
  data = NULL,
  geom = "line",
  stat = "identity",
  position = "identity",
  ...,
  n = 101,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use display the data}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{extra arguments, which go partly into the geom/stat parameters, and partly into extra_params (which go into the `predict` call)}

\item{n}{Specifies whether the prediction must be performed on interpolated points along the x-axis.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
You can use this just like a regular `layer` call.
Beware that this function does not warn you about spurious or misspelled arguments, as
these will be passed along to `predict()` instead.
}
\examples{
m1 <- tdmore(theopp_nlmixr)
db <- dataTibble(
     observed=data.frame(ID=c(1,2), TIME=5, CONC=c(10,15)),
     regimen=data.frame(TIME=0, AMT=5)
)
pred <- posthoc(object=m1, db)
ggplot2::ggplot(pred$fit[[1]], ggplot2::aes(x=TIME, y=CONC)) +
    ggplot2::geom_point() + predictionLayer(newdata=seq(0, 10, 0.1)) +
    predictionLayer(data=pred$fit[[2]], newdata=seq(0, 10, 0.1), color=2)
}
