% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommendation.R
\name{print.recommendation}
\alias{print.recommendation}
\title{Print a recommendation object.}
\usage{
\method{print}{recommendation}(x, ...)
}
\arguments{
\item{x}{a recommendation object}

\item{...}{ignored}
}
\description{
Print a recommendation object.
}
