% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{print.summary.tdmore}
\alias{print.summary.tdmore}
\title{Print a tdmore summary}
\usage{
\method{print}{summary.tdmore}(x, ...)
}
\arguments{
\item{x}{a tdmore summary}
}
\description{
Print a tdmore summary
}
