% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc.R
\name{print.tdmore_mpc}
\alias{print.tdmore_mpc}
\title{Print a tdmore_mpc object}
\usage{
\method{print}{tdmore_mpc}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{ignored}
}
\description{
Print a tdmore_mpc object
}
