% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMixture.R
\name{print.tdmoreCurried}
\alias{print.tdmoreCurried}
\title{Print a tdmoreCurried object.}
\usage{
\method{print}{tdmoreCurried}(x, ...)
}
\arguments{
\item{x}{a tdmoreCurried object}

\item{...}{extra arguments, not used}
}
\description{
Print a tdmoreCurried object.
}
