% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{processParameters}
\alias{processParameters}
\title{Check input parameters and return a standardised named num array with the parameter names and values.
If the initial values are not provided, zeroes will be used (=population values for ETAS)
Provided parameters will overwrite initial values.}
\usage{
processParameters(parameters, tdmore, regimen, initialValues = NULL)
}
\arguments{
\item{parameters}{user inputted array of parameters}

\item{tdmore}{the tdmore object}

\item{regimen}{the regimen}

\item{initialValues}{the initial values, in the right order}
}
\value{
a standardised named num array with the parameter names and values
}
\description{
Check input parameters and return a standardised named num array with the parameter names and values.
If the initial values are not provided, zeroes will be used (=population values for ETAS)
Provided parameters will overwrite initial values.
}
