% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{profile.tdmorefit}
\alias{profile.tdmorefit}
\title{Get an overview of the log-likelihood for varying parameter values.
If arguments `fix` and `limits` are not set for a specific parameter, individual eta's will be generated to cover 95 percent of the population distribution.}
\usage{
profile.tdmorefit(fitted, fix = NULL, maxpts = 50, limits = NULL,
  type = c("ll", "pop", "pred"), .progress = "none", ...)
}
\arguments{
\item{fitted}{A tdmorefit object}

\item{fix}{Which parameters to fix? Named vector of the parameters that are fixed and should not be profiled}

\item{maxpts}{Maximum number of points per parameter}

\item{limits}{limits to explore (numeric vector of form c(min, max) or specific limits per parameter in the form list(ETA1=c(min,max), ETA2=c(min,max), etc))}

\item{type}{log-lokelihood function type, 3 possible values: 'pop', 'pred' or 'll' (= pop + pred)}

\item{.progress}{See plyr::ddply}

\item{...}{ignored}
}
\value{
a tdmore profile object. It namely contains a data.frame with each parameter value tested, and an additional `logLik` column with the log-likelihood for each parameter combination.
}
\description{
Get an overview of the log-likelihood for varying parameter values.
If arguments `fix` and `limits` are not set for a specific parameter, individual eta's will be generated to cover 95 percent of the population distribution.
}
