% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proseval.R
\name{proseval}
\alias{proseval}
\title{This function calculates a prospective evaluation.
For every ID, it will calculate as many fits as there are rows of observations +1.
The first fit will be the population fit (not using any observations).
The next fit will only use one observation.
The next fit will use an extra observation.
And so on.}
\usage{
proseval(tdmore, observed = NULL, regimen = NULL, covariates = NULL,
  par = NULL, fix = NULL, ...)
}
\arguments{
\item{observed}{data frame with at least a TIME column, and all observed data. The observed data will be compared to the model predictions.
If not specified, we estimate the population prediction}

\item{regimen}{data frame describing the treatment regimen.}

\item{covariates}{the model covariates}

\item{par}{optional starting parameter for the MLE minimization}

\item{fix}{named vector with the fixed parameters}

\item{...}{passed to \link{estimate} function}
}
\value{
a tibble with columns ID,
OBS (showing how many observations were included in this fit),
and column `fit` with the tdmorefit object (nested in a list)
}
\description{
This function calculates a prospective evaluation.
For every ID, it will calculate as many fits as there are rows of observations +1.
The first fit will be the population fit (not using any observations).
The next fit will only use one observation.
The next fit will use an extra observation.
And so on.
}
