% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R, R/model.R
\name{residuals.tdmorefit}
\alias{residuals.tdmorefit}
\alias{residuals.tdmore}
\title{Get the residuals for a specific tdmorefit}
\usage{
\method{residuals}{tdmorefit}(object, data, weighted = FALSE, ...)

\method{residuals}{tdmore}(object, predicted, observed, weighted = FALSE,
  inverse = FALSE, ...)
}
\arguments{
\item{object}{A TDMore object}

\item{data}{optional data.frame
If specified, this adds the original residuals back onto the data.frame
If points at a time without corresponding IRES are present, an IRES of 0 is used and a warning is emitted.}

\item{weighted}{Should the residuals be divided by the standard deviation of the error model?}

\item{...}{ignored}

\item{predicted}{What was predicted, as data.frame}

\item{observed}{The observed values, as data.frame}

\item{inverse}{if TRUE, perform the inverse operation (assume `predicted` has IPRED values, and `observed` has IRES/IWRES values)}
}
\value{
The `predicted` data.frame, with all output values replaced by IRES or IWRES
}
\description{
Get the residuals for a specific tdmorefit

Get the residual values of a predicted value vs the observed value
}
\examples{
observed <- data.frame(TIME=5, CONC=0.060)
regimen=data.frame(TIME=0, AMT=10)
fit <- estimate(tdmore(default_model),
    regimen=regimen,
    observed=observed,
    covariates=c(WT=70)
    )
wres <- residuals(fit, weighted=TRUE)

all.equal(
  residuals(fit, predict(fit), weighted=TRUE),
  observed
)

regimen <- fit$regimen
regimen$AMT <- regimen$AMT*2
predictionForDoubleDose <- residuals(fit, predict(fit, regimen=regimen), weighted=TRUE)

}
