% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{residuals.tdmore}
\alias{residuals.tdmore}
\title{Get the residual values of a predicted value vs the observed value}
\usage{
\method{residuals}{tdmore}(object, observed, predicted, weighted = FALSE,
  ...)
}
\arguments{
\item{object}{A TDMore object}

\item{observed}{The observed values, as data.frame}

\item{predicted}{What was predicted, as data.frame}

\item{weighted}{Should the residuals be divided by the standard deviation of the error model?}

\item{...}{ignored}
}
\value{
A numeric vector, equivalent to IRES or IWRES
}
\description{
Get the residual values of a predicted value vs the observed value
}
