% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{roundTime}
\alias{roundTime}
\title{For a given set of planned times, round to the nearest reference time.}
\usage{
roundTime(x, reference, range = c(-0.5, 0.5), if.na = c("keep", "NA"))
}
\arguments{
\item{x}{a vector of times to round}

\item{reference}{a vector of times, used to round}

\item{range}{how much can be subtracted or added from/to 'x' to get to the reference}

\item{if.na}{if no corresponding reference time is found, what should we do?
if 'keep', then the original time is kept
if 'NA', then we set the value to NA}
}
\value{
A vector of the same length as x. Values close to reference values are replaced by the reference value.
}
\description{
This can be used to move an intended administration to fit the historic treatment regimen of a
patient in retrospective data.
}
