% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{sampleMC}
\alias{sampleMC}
\title{Generate parameter samples using uncertainty around the fit.}
\usage{
sampleMC(
  tdmorefit,
  fix = tdmorefit$fix,
  mc.maxpts = 100,
  .progress = interactive()
)
}
\arguments{
\item{tdmorefit}{a tdmorefit object}

\item{fix}{named numeric vector with the fixed parameters}

\item{mc.maxpts}{number of Monte-Carlo samples}

\item{.progress}{Allows to specify a plyr-like progress object
A plyr progress object is a list with 3 function definitions: `init(N)`, `step()` and `term()`.
This can also be specified as a boolean. TRUE uses the default dplyr progress_estimated.}
}
\value{
the Monte-Carlo matrix, first column is always the 'sample' column
}
\description{
If a variance-covariance matrix could be generated during estimation using asymptotic , a simple multivariate normal sampling is used.
If the variance-covariance matrix is NULL, we use Metropolis-Hastings sampling.
}
