% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{sampleMC_metrop}
\alias{sampleMC_metrop}
\title{Generate parameter samples using the Metropolis MCMC method.}
\usage{
sampleMC_metrop(tdmorefit, fix = tdmorefit$fix, mc.maxpts = 100, verbose = 0)
}
\arguments{
\item{tdmorefit}{a tdmorefit object}

\item{fix}{named numeric vector with the fixed parameters}

\item{mc.maxpts}{number of Monte-Carlo samples}

\item{verbose}{A switch which determines whether or not the progress of the
sampler is printed to the screen.  If \code{verbose} is greater than 0 the
iteration number, the \eqn{\theta} vector, the function value, and
the Metropolis acceptance rate are sent to the screen every \code{verbose}th
iteration.}
}
\value{
the Monte-Carlo matrix, first column is always the 'sample' column
}
\description{
Generate parameter samples using the Metropolis MCMC method.
}
