% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{sampleMC_norm}
\alias{sampleMC_norm}
\title{Generate parameter samples using a muilti-variate normal distribution around the variance-covariance matrix}
\usage{
sampleMC_norm(tdmorefit, fix = tdmorefit$fix, mc.maxpts = 100)
}
\arguments{
\item{tdmorefit}{a tdmorefit object}

\item{fix}{named numeric vector with the fixed parameters}

\item{mc.maxpts}{number of Monte-Carlo samples}
}
\value{
the Monte-Carlo matrix, first column is always the 'sample' column
}
\description{
Generate parameter samples using a muilti-variate normal distribution around the variance-covariance matrix
}
