% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc.R
\name{selectBestCovariates}
\alias{selectBestCovariates}
\title{Select most appropriate covariates.}
\usage{
selectBestCovariates(t, covariates, thetaNames)
}
\arguments{
\item{t}{time}

\item{covariates}{time-varying covariates}

\item{thetaNames}{theta names to be removed}
}
\value{
a vector with the covariates
}
\description{
Select most appropriate covariates.
}
