% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyProfile.R
\name{shinyProfile}
\alias{shinyProfile}
\title{Open a Shiny Gadget to explore the log-likelihood profile
of a specific fit}
\usage{
shinyProfile(tdmorefit, fix = NULL, ...)
}
\description{
Open a Shiny Gadget to explore the log-likelihood profile
of a specific fit
}
