% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyProfile.R
\name{shinyProfileApp}
\alias{shinyProfileApp}
\title{Build a Shiny app to explore the log-likelihood profile
of a specific fit}
\usage{
shinyProfileApp(fitted, fix = NULL, ...)
}
\arguments{
\item{fitted}{A tdmorefit object}

\item{fix}{Which parameters to fix? Named vector of the parameters that are fixed and should not be profiled
Specify a FIX of NA to estimate the optimal value at that specific grid-point.}

\item{...}{Extra arguments passed to `profile`}
}
\description{
Build a Shiny app to explore the log-likelihood profile
of a specific fit
}
