% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{ShinyToDplyrProgressFacade}
\alias{ShinyToDplyrProgressFacade}
\title{A facade object to use Shiny Progress with dplyr}
\description{
Used to translate dplyr progress to a Shiny progress bar
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{proxy}}{Shiny Progress object}

\item{\code{n}}{total ticks for dplyr progress; can be reinitialized}

\item{\code{i}}{current tick for dplyr progress; can be reinitialized}

\item{\code{offset}}{offset for Shiny Progress}

\item{\code{amount}}{total advanced amount in Shiny after the dplyr progress is done}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ShinyToDplyrProgressFacade$new()}}
\item \href{#method-tick}{\code{ShinyToDplyrProgressFacade$tick()}}
\item \href{#method-stop}{\code{ShinyToDplyrProgressFacade$stop()}}
\item \href{#method-print}{\code{ShinyToDplyrProgressFacade$print()}}
\item \href{#method-clone}{\code{ShinyToDplyrProgressFacade$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialize a shiny progress to dplyr facade. This can either be called with the `proxy` and `amount`
arguments for a first initialization (e.g. when using `new`), or with
`n` and `min_time` arguments for the dplyr-like initialization
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyToDplyrProgressFacade$new(
  n = 1,
  min_time = 0,
  proxy = NULL,
  amount = 1,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Total number of items}

\item{\code{min_time}}{Progress bar will wait until at least `min_time`
seconds have elapsed before displaying any results.}

\item{\code{proxy}}{Shiny Progress object}

\item{\code{amount}}{amount advanced in Shiny Progress}

\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an object that can be used in `.progress` arguments for dplyr
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tick"></a>}}
\subsection{Method \code{tick()}}{
Advance a tick
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyToDplyrProgressFacade$tick()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop"></a>}}
\subsection{Method \code{stop()}}{
Stop the timer; advance until the end
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyToDplyrProgressFacade$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print out the progress; transfers the progress to Shiny
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyToDplyrProgressFacade$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyToDplyrProgressFacade$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
