% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{ShinyToProgressFacade}
\alias{ShinyToProgressFacade}
\title{A facade object to use Shiny Progress as if they are a progress::progress_bar}
\description{
Used to translate dplyr progress to a Shiny progress bar
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{proxy}}{Shiny Progress object}

\item{\code{total}}{total ticks for progress; can be reinitialized}

\item{\code{i}}{current tick for progress; can be reinitialized}

\item{\code{offset}}{offset for Shiny Progress}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ShinyToProgressFacade$new()}}
\item \href{#method-tick}{\code{ShinyToProgressFacade$tick()}}
\item \href{#method-clone}{\code{ShinyToProgressFacade$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a shiny progress to dplyr facade. This can either be called with the `proxy` and `amount`
arguments for a first initialization (e.g. when using `new`), or with
`n` and `min_time` arguments for the dplyr-like initialization
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyToProgressFacade$new(total = 1, proxy = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{total}}{Total number of items}

\item{\code{proxy}}{Shiny Progress object}

\item{\code{...}}{ignored}

\item{\code{min_time}}{Progress bar will wait until at least `min_time`
seconds have elapsed before displaying any results.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an object that can be used in `.progress` arguments for dplyr
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tick"></a>}}
\if{latex}{\out{\hypertarget{method-tick}{}}}
\subsection{Method \code{tick()}}{
Advance a tick
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyToProgressFacade$tick()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyToProgressFacade$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
