% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{summariseFittedMC}
\alias{summariseFittedMC}
\title{Summarise fitted MC matrix over all the samples.
TIME column must be present.}
\usage{
summariseFittedMC(fittedMC, ipred, level, oNames)
}
\arguments{
\item{fittedMC}{fitted MC matrix (result of predict.tdmorefit)}

\item{ipred}{result of predict.tdmorefit}

\item{level}{ignored}

\item{oNames}{outputs to summarise}
}
\value{
a summarised data frame
}
\description{
Summarise fitted MC matrix over all the samples.
TIME column must be present.
}
