% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommendation.R
\name{summary.recommendation}
\alias{summary.recommendation}
\title{Summarize a recommendation object.}
\usage{
\method{summary}{recommendation}(object, ...)
}
\arguments{
\item{object}{a recommendation object}

\item{...}{ignored}
}
\description{
Summarize a recommendation object.
}
