% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{summary.tdmore}
\alias{summary.tdmore}
\title{Summarise a tdmore object.}
\usage{
\method{summary}{tdmore}(object, ...)
}
\arguments{
\item{object}{a tdmore object}

\item{...}{ignored}
}
\description{
Summarise a tdmore object.
}
