% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{summary.tdmorefit}
\alias{summary.tdmorefit}
\title{Summary of an EBE fit}
\usage{
\method{summary}{tdmorefit}(object, ...)
}
\arguments{
\item{object}{A tdmorefit object}

\item{...}{Additional parameters ignored}
}
\value{
Nothing
}
\description{
Summary of an EBE fit
}
