% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-models.R
\docType{data}
\name{tacrolimus_storset}
\alias{tacrolimus_storset}
\title{NlmixrUI object describing the Tacrolimus model from Størset, Elisabet, et al.}
\format{an nlmixrUI model}
\usage{
tacrolimus_storset
}
\description{
NlmixrUI object describing the Tacrolimus model from Størset, Elisabet, et al.
}
\keyword{datasets}
