% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{target}
\alias{target}
\title{Create a new target.}
\usage{
target(min, max)
}
\arguments{
\item{min}{the min value of the target, same unit as the one defined in output}

\item{max}{the max value of the target, same unit as the one defined in output}
}
\value{
a tdmore_target object
}
\description{
Create a new target.
}
\seealso{
Other metadata: 
\code{\link{covariate}()},
\code{\link{formulation}()},
\code{\link{getDosingInterval}()},
\code{\link{getMetadataByClass}()},
\code{\link{getMetadataByName}()},
\code{\link{metadata}()},
\code{\link{output}()}
}
\concept{metadata}
