% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmore-package.R
\docType{package}
\name{tdmore-package}
\alias{tdmore-package}
\title{TDMore: Estimate individual parameters for population pharmacometrics models, and use them to make predictions and/or dose recommendations.}
\description{
TDMore: Estimate individual parameters for population pharmacometrics models, and use them to make predictions and/or dose recommendations.
}
