% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraic.R
\name{tdmore.algebraic}
\alias{tdmore.algebraic}
\title{Build a tdmore object based on an algebraic model}
\usage{
\method{tdmore}{algebraic}(model, res_var, parameters = NULL,
  omega = NULL, ...)
}
\arguments{
\item{model}{The algebraic model}

\item{res_var}{the residual variability}

\item{parameters}{List of parameters, should be NULL}

\item{omega}{the omega matrix of the model}

\item{...}{extra arguments will be passed to the model_predict call}
}
\value{
a tdmore object, capable of estimating bayesian individual parameters
}
\description{
Build a tdmore object based on an algebraic model
}
