% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{tdmore.default}
\alias{tdmore.default}
\title{Default function, catching unsupported types of models}
\usage{
\method{tdmore}{default}(model, ...)
}
\arguments{
\item{model}{model object}

\item{...}{ignored}
}
\description{
Default function, catching unsupported types of models
}
