% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr.R
\name{tdmore.nlmixrFitCore}
\alias{tdmore.nlmixrFitCore}
\title{Create a TDM-capable model from an nlmixr result.}
\usage{
\method{tdmore}{nlmixrFitCore}(model, ...)
}
\arguments{
\item{model}{the fitted nlmixr model}

\item{...}{extra arguments will be passed to the model_predict call}
}
\value{
An object of class tdmore, which can be used to estimate posthoc Bayesian parameters.
The model contained within is actually an RxODE object.
}
\description{
Create a TDM-capable model from an nlmixr result.
}
