% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr.R
\name{tdmore.nlmixrUI}
\alias{tdmore.nlmixrUI}
\title{Create a TDM-capable model from an nlmixr UI object.}
\usage{
\method{tdmore}{nlmixrUI}(model, iov = NULL, ...)
}
\arguments{
\item{model}{the nlmixr UI object}

\item{iov}{list of parameter names related to IOV, NULL if no IOV}

\item{...}{extra arguments will be passed to the model_predict call}
}
\value{
An object of class tdmore, which can be used to estimate posthoc Bayesian parameters.
The model contained within is actually an RxODE object.
}
\description{
Create a TDM-capable model from an nlmixr UI object.
}
