% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R, R/tdmore-package.R
\docType{package}
\name{tdmore}
\alias{tdmore}
\alias{tdmore-package}
\title{tdmore is a generic function to append TDM functionality to a pharmacometrics model.}
\usage{
tdmore(model, ...)
}
\arguments{
\item{model}{the base model}

\item{...}{extra arguments}
}
\value{
An object of class tdmore, which can be used to estimate posthoc Bayesian parameters
}
\description{
In some cases, this function will automatically guess the required parameters from the model specified (e.g. names of parameters, names of covariates, residual error model).
In other cases, you will have to specify this manually.

Estimate individual parameters for population pharmacometrics models, and use them to make predictions and/or dose recommendations.
}
