% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{tdmore.RxODE}
\alias{tdmore.RxODE}
\title{Create a TDM-capable model from an RxODE model}
\usage{
\method{tdmore}{RxODE}(model, res_var, parameters = NULL, omega = NULL,
  ...)
}
\arguments{
\item{model}{the RxODE model}

\item{res_var}{the residual variability}

\item{parameters}{list of parameter names, or NULL to use all parameters names from RxODE}

\item{omega}{omega variance-covariance matrix, or NULL to use a diagonal matrix of variances 1}

\item{...}{extra arguments will be passed to the model_predict call}
}
\value{
An object of class tdmore, which can be used to estimate posthoc Bayesian parameters
}
\description{
Create a TDM-capable model from an RxODE model
}
\examples{
modelCode <- "
CL = 23.6 * exp(ETA1*0.42);
Vc = 1070 * exp(ETA2*1.11);
ka=4.48;
CONC = centr / Vc * 1000;
d/dt(abs) = -ka*abs;
d/dt(centr) = ka*abs - CL/Vc*centr;
"
rxodeModel <- RxODE::RxODE(modelCode)
model <- tdmore(rxodeModel, res_var=list(errorModel(var="CONC", add=3.7)))
}
