% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMixture.R
\name{tdmore_mixture}
\alias{tdmore_mixture}
\title{Create a TDMore mixture model.}
\usage{
tdmore_mixture(..., probs)
}
\arguments{
\item{...}{1 or more tdmore models, to be added in the set.}

\item{probs}{a numeric vector with the model probabilities. Sum must be 1.}
}
\value{
a tdmore_set object
}
\description{
Create a TDMore mixture model.
}
