% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSet.R
\name{tdmore_set}
\alias{tdmore_set}
\title{Create a TDMore set.
Note that the order of the given models matters.
Order must be defined from the most restrictive model (with a high number of covariates) to the lowest restrictive model (with a low number of covariates).
These models will be tried one by one from left to right with the 'observed' data.
The first model matching the covariates from the 'observed' data will be used.}
\usage{
tdmore_set(...)
}
\arguments{
\item{...}{1 or more TDMore or TDMore mixture models, to be added in the set.}
}
\value{
a tdmore_set object
}
\description{
Create a TDMore set.
Note that the order of the given models matters.
Order must be defined from the most restrictive model (with a high number of covariates) to the lowest restrictive model (with a low number of covariates).
These models will be tried one by one from left to right with the 'observed' data.
The first model matching the covariates from the 'observed' data will be used.
}
