% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{tdmorefit}
\alias{tdmorefit}
\title{Create a tdmorefit object manually}
\usage{
tdmorefit(
  tdmore,
  observed = NULL,
  regimen,
  covariates = NULL,
  ofv = NA,
  res = NULL,
  varcov = NULL,
  fix = NULL,
  nlmresult = NULL,
  call = NULL
)
}
\arguments{
\item{tdmore}{the tdmore object}

\item{observed}{the observed data.frame, or NULL}

\item{regimen}{the treatment regimen data.frame}

\item{covariates}{the model covariates}

\item{ofv}{(optional) the OFV value}

\item{res}{the found parameter values, as a named vector, or NULL to use 0.}

\item{varcov}{the found varcov matrix}

\item{fix}{the fixed parameters}

\item{nlmresult}{the result of the non-linear minimization}

\item{call}{a informative string that gives the arguments that were used in estimate()}
}
\value{
A tdmorefit object, manually created
}
\description{
Create a tdmorefit object manually
}
