% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelLibrary.R
\docType{data}
\name{theopp}
\alias{theopp}
\title{Theophylline dataset, as included in the NONMEM distribution.}
\format{
A data frame with 12x13 rows: 1 administration and 12 observation rows
\describe{
  \item{ID}{Subject ID}
  \item{AMT}{Theophylline dose, in mg}
  \item{TIME}{Time of observation, in days after first dose}
  \item{CONC}{Concentration, in mg/L}
  \item{EVID}{0 for observation, 1 for administration}
  \item{WT}{Weight in kg}
}
}
\usage{
theopp
}
\description{
This is comprised of 12 subjects with each 12 concentrations.
}
\keyword{datasets}
