% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelLibrary.R
\docType{data}
\name{theopp_nlmixr}
\alias{theopp_nlmixr}
\title{NlmixrUI object describing theophylline 1-compartment model with oral absorption}
\format{
A fitted Nlmixr model
}
\usage{
theopp_nlmixr
}
\description{
The model was estimated by nlmixr SAEM
}
\keyword{datasets}
