% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.autolayer.R
\name{timeRange}
\alias{timeRange}
\title{Determine a timerange suitable for plotting each aspect of a tdmorefit}
\usage{
timeRange(ipred = NULL, regimen = NULL, observed = NULL, N = 101)
}
\arguments{
\item{ipred}{tdmorefit object to calculate the time range}

\item{regimen}{the regimen, or not specified to use the original regimen}

\item{observed}{observed values, or not specified to use the original observed values}

\item{N}{number of interpolation points}
}
\description{
Determine a timerange suitable for plotting each aspect of a tdmorefit
}
