% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{toString.tdmore_covariate}
\alias{toString.tdmore_covariate}
\title{Tdmore_covariate to string method.}
\usage{
\method{toString}{tdmore_covariate}(x, ...)
}
\arguments{
\item{x}{a tdmore_covariate object}

\item{...}{ignored}
}
\description{
Tdmore_covariate to string method.
}
