% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{toString.tdmore_dose}
\alias{toString.tdmore_dose}
\title{Tdmore_dose to string method.}
\usage{
\method{toString}{tdmore_dose}(x, ...)
}
\arguments{
\item{x}{a tdmore object}

\item{...}{ignored}
}
\description{
Tdmore_dose to string method.
}
