% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{toString.tdmore_formulation}
\alias{toString.tdmore_formulation}
\title{Tdmore_formulation to string method.}
\usage{
\method{toString}{tdmore_formulation}(x, ...)
}
\arguments{
\item{x}{a tdmore_dose object}

\item{...}{ignored}
}
\description{
Tdmore_formulation to string method.
}
