% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{toString.tdmore_output}
\alias{toString.tdmore_output}
\title{Tdmore_output to string method.}
\usage{
\method{toString}{tdmore_output}(x, ...)
}
\arguments{
\item{x}{a tdmore object}

\item{...}{ignored}
}
\description{
Tdmore_output to string method.
}
