% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{toString.tdmore_target}
\alias{toString.tdmore_target}
\title{Tdmore_target to string method.}
\usage{
\method{toString}{tdmore_target}(x, ...)
}
\arguments{
\item{x}{a tdmore_target object}

\item{...}{ignored}
}
\description{
Tdmore_target to string method.
}
