% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSpace.R
\name{transformRegimen}
\alias{transformRegimen}
\title{Transform a specific treatment regimen using a modifier}
\usage{
transformRegimen(regimen, x)
}
\arguments{
\item{regimen}{a treatment regimen}

\item{x}{A row of a possibilitygrid}
}
\value{
The modified treatment regimen
}
\description{
Transform a specific treatment regimen using a modifier
}
