% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelLibrary.R
\docType{data}
\name{two_error_models}
\alias{two_error_models}
\title{Default example model
TODO: Better documentation required}
\format{an nlmixrUI model}
\usage{
two_error_models
}
\description{
Default example model
TODO: Better documentation required
}
\keyword{datasets}
