% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.qdoptimize.R
\name{updateRegimen}
\alias{updateRegimen}
\title{Update a regimen with the specified dose.}
\usage{
updateRegimen(regimen, doseRows = NULL, newDose)
}
\arguments{
\item{regimen}{the regimen to update}

\item{doseRows}{which rows of the regimen to adapt, if not specified, the last dose will be adapted}

\item{newDose}{the specified new dose}
}
\value{
the updated regimen
}
\description{
Update a regimen with the specified dose.
}
