% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{vcov.tdmorefit}
\alias{vcov.tdmorefit}
\title{The variance-covariance matrix}
\usage{
\method{vcov}{tdmorefit}(object, ...)
}
\arguments{
\item{object}{A tdmorefit object}

\item{...}{ignored}
}
\value{
Named numeric matrix
}
\description{
The variance-covariance matrix
}
