% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{vectorToDiagonalMatrix}
\alias{vectorToDiagonalMatrix}
\title{Convert a numeric vector/list into a diagonal matrix.
If the vector/list is named, names will be used as the column and row names of the resulting diagonal matrix.}
\usage{
vectorToDiagonalMatrix(vector)
}
\arguments{
\item{vector}{a numeric vector/list}
}
\value{
a diagonal matrix
}
\description{
Convert a numeric vector/list into a diagonal matrix.
If the vector/list is named, names will be used as the column and row names of the resulting diagonal matrix.
}
\keyword{internal}
