% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc.R
\name{addIterationColumn}
\alias{addIterationColumn}
\title{Add an iteration column 'ITER' to the observed dataframe.
Observations will be grouped by occasions.}
\usage{
addIterationColumn(regimen, observed)
}
\arguments{
\item{regimen}{regimen}

\item{observed}{observed}
}
\value{
modified observed dataframe, with an 'ITER' column
}
\description{
Add an iteration column 'ITER' to the observed dataframe.
Observations will be grouped by occasions.
}
