% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{addRegimenToEventTable}
\alias{addRegimenToEventTable}
\title{Add the given regimen into the RxODE event table.}
\usage{
addRegimenToEventTable(eventTable, regimen, nbSSDoses = NULL)
}
\arguments{
\item{eventTable}{the RxODE event table}

\item{regimen}{the specified regimen}

\item{nbSSDoses}{number of doses to prepend for simulation of steady-state}
}
\value{
a completed event table
}
\description{
Add the given regimen into the RxODE event table.
}
