% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraic.R
\name{algebraic}
\alias{algebraic}
\title{Use an external function to generate an algebraic model, capable of being used with tdmore()}
\usage{
algebraic(algebraicDefinition)
}
\arguments{
\item{algebraicDefinition}{the algebraic model definition}
}
\value{
An algebraic model
}
\description{
Use an external function to generate an algebraic model, capable of being used with tdmore()
}
