% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraic.R
\name{algebraic}
\alias{algebraic}
\title{Use an external function to generate an algebraic model, capable of being used with tdmore()
Algebraic functions are assumed to be dose-linear, meaning that they can be calculated for a single dose
and then summed up for multiple doses.}
\usage{
algebraic(fun, output = "CONC")
}
\arguments{
\item{fun}{a function with 'time' as the first argument, and additional parameters as needed.
In case parameters will be estimated, they should have a mean of '0' and be normal-distributed.

The function can also include parameters `TIME`, `AMT`, `II`, `ADDL`, `RATE`, `DURATION`, `CMT`.
These are then taken from the treatment regimen.}

\item{output}{Name for the output}
}
\value{
An algebraic prediction model
}
\description{
Use an external function to generate an algebraic model, capable of being used with tdmore()
Algebraic functions are assumed to be dose-linear, meaning that they can be calculated for a single dose
and then summed up for multiple doses.
}
