% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{as.population}
\alias{as.population}
\title{Create a typical value subject}
\usage{
as.population(x, ...)
}
\arguments{
\item{x}{a tdmorefit object, or tdmore object}

\item{...}{extra parameters passed to the internal engine}
}
\description{
Create a typical value subject
}
