% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.autolayer.R
\name{autoplot.tdmore}
\alias{autoplot.tdmore}
\title{Plot a tdmore object as the population}
\usage{
\method{autoplot}{tdmore}(x, regimen = NULL, covariates = NULL, newdata = NULL, ...)
}
\arguments{
\item{x}{a tdmore object}

\item{regimen}{the regimen}

\item{covariates}{the covariates}

\item{newdata}{passed to `predict`}

\item{...}{extra arguments}
}
\description{
Plot a tdmore object as the population
}
