% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.autolayer.R
\name{autoplot.tdmore}
\alias{autoplot.tdmore}
\alias{autolayer.tdmorefit}
\title{Automatically plot a tdmorefit object}
\usage{
\method{autolayer}{tdmorefit}(
  x,
  color = "tomato1",
  fill = color,
  alpha = 0.2,
  se.fit = T,
  level = 0.95,
  ...,
  na.rm = TRUE
)

\method{autoplot}{tdmore}(x, regimen = NULL, covariates = NULL, newdata = NULL, ...)
}
\arguments{
\item{x}{a tdmore object}

\item{color}{color to use for the line}

\item{fill}{fill color to use for the ribbon}

\item{alpha}{alpha level for the ribbon}

\item{se.fit}{plot a confidence band around the prediction?}

\item{level}{level for the confidence band}

\item{...}{extra arguments}

\item{na.rm}{should NA be removed}

\item{regimen}{the regimen}

\item{covariates}{the covariates}

\item{newdata}{passed to `predict`}
}
\description{
The variable plotted on the y-axis is the first column with residual error that is available in the predicted result.
You can easily influence this by providing a `newdata` argument.
}
