% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.tdmorefit}
\alias{autoplot.tdmorefit}
\title{Plot a tdmorefit object.}
\usage{
\method{autoplot}{tdmorefit}(x, newdata = NULL, ...)
}
\arguments{
\item{x}{the tdmorefit object}

\item{newdata}{a data.frame with at least TIME and any other columns to plot, NULL to plot all columns from the original observed data between time 0 and max(observationTime) or a numeric vector of times}

\item{...}{ignored}
}
\value{
a ggplot object with the fitted individual curve, the 95% ci of this curve, the population prediction, the observed data and the residual error around the observed data
}
\description{
Plot a tdmorefit object.
}
