% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.autolayer.R
\name{autoplot.tdmorefit}
\alias{autoplot.tdmorefit}
\title{Plot a tdmorefit object.}
\usage{
\method{autoplot}{tdmorefit}(x, population = TRUE, fit = TRUE, ...)
}
\arguments{
\item{x}{the tdmorefit object}

\item{population}{should I plot the population prediction (in blue) ?}

\item{fit}{should I plot the fit (in red) ?}

\item{...}{passed on to `predict`}
}
\value{
a ggplot object with the fitted individual curve, the 95% CI of this curve, the population prediction (with between-subject variability) and the observed data
}
\description{
Plot a tdmorefit object.
}
