% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.qdoptimize.R
\name{convertMCResultToRecommendation}
\alias{convertMCResultToRecommendation}
\title{Convert the Monte-Carlo uniroot results into a recommendation object.}
\usage{
convertMCResultToRecommendation(mcResult, regimen, doseRows, level)
}
\arguments{
\item{mcResult}{a dataframe with all the uniroot results, one per Monte-Carlo sample}

\item{regimen}{the regimen}

\item{doseRows}{which rows of the regimen to adapt when searching for a new dose, or NULL to take the last one}

\item{level}{the confidence interval on the dose}
}
\value{
a recommendation object
}
\description{
Convert the Monte-Carlo uniroot results into a recommendation object.
}
\keyword{internal}
