% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMixture.R
\name{curry_covariate}
\alias{curry_covariate}
\title{Fill in a specific covariate in a model, transforming it into a more specific model.}
\usage{
curry_covariate(model, curriedCovariates)
}
\arguments{
\item{model}{tdmore model with a discrete covariate}

\item{curriedCovariates}{the curried covariates}
}
\value{
the model with the curried covariates
}
\description{
Fill in a specific covariate in a model, transforming it into a more specific model.
}
