% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorModel.R
\name{errorModel}
\alias{errorModel}
\title{Instantiate a new error model.}
\usage{
errorModel(var = "CONC", add = 0, prop = 0, type = NULL)
}
\arguments{
\item{var}{the related variable name}

\item{add}{additive residual error, as stdev}

\item{prop}{proportional residual error, as stdev}

\item{type}{type of error: constant, proportional, combined1, combined2, exponential
or NULL to auto-detect. A combination of additive and proportional always assumes 'combined2'.}
}
\description{
Instantiate a new error model.
}
\details{
Please note that convention differ between modeling software. While NONMEM would always code a
combined additive/proportional model as `W=SQRT( SIGMA(1)**2 + SIGMA(2)**2 * IPRED**2 )`,
Monolix uses a default of `CONC = Cc + (a + b*Cc)*e`. TDMore uses the NONMEM convention by default.
}
