% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{estimate}
\alias{estimate}
\title{Calculate the Empirical Bayesian Estimate parameters that predict
a given dataset using the model}
\usage{
estimate(tdmore, observed = NULL, regimen, covariates = NULL,
  par = NULL, method = "L-BFGS-B", lower = -5 *
  sqrt(diag(tdmore$omega)), upper = +5 * sqrt(diag(tdmore$omega)), ...)
}
\arguments{
\item{tdmore}{a tdmore object}

\item{observed}{data frame with at least a TIME column, and all observed data. The observed data will be compared to the model predictions.
If not specified, we estimate the population prediction}

\item{regimen}{data frame describing the treatment regimen.}

\item{covariates}{the model covariates}

\item{par}{optional starting parameter for the MLE minimization}

\item{method}{the optimisation method, by default, method "L-BFGS-B" is used}

\item{lower}{the lower bounds of the parameters, by default, -5 * sqrt(diag(tdmore$omega)) is used}

\item{upper}{the upper bounds of the parameters, by default, +5 * sqrt(diag(tdmore$omega)) is used}

\item{...}{extra parameters to pass to the optim function}
}
\value{
A tdmorefit object
}
\description{
Calculate the Empirical Bayesian Estimate parameters that predict
a given dataset using the model
}
