% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMixture.R
\name{estimateMixtureModel}
\alias{estimateMixtureModel}
\title{Calculate the Empirical Bayesian Estimate parameters that predict
a given dataset using a mixture model.}
\usage{
estimateMixtureModel(
  object,
  observed = NULL,
  regimen,
  covariates = NULL,
  par = NULL,
  fix = NULL,
  method = "L-BFGS-B",
  lower = NULL,
  upper = NULL,
  ...
)
}
\arguments{
\item{object}{a tdmore, tdmore_set or tdmore_mixture object}

\item{observed}{data frame with at least a TIME column, and all observed data. The observed data will be compared to the model predictions.
If not specified, we estimate the population prediction}

\item{regimen}{data frame describing the treatment regimen.}

\item{covariates}{the model covariates}

\item{par}{optional starting parameter for the MLE minimization}

\item{fix}{named numeric vector with the fixed parameters. Names can repeat to fix parameters in multiple occasions.}

\item{method}{the optimisation method, by default, method "L-BFGS-B" is used
Can also be specified as a list, in which case all methods specified will be tried and the one finding the estimate with highest log-likelihood will be returned.}

\item{lower}{the lower bounds of the parameters, if null, -5 * sqrt(diag(model$omega)) is used}

\item{upper}{the upper bounds of the parameters, if null, +5 * sqrt(diag(model$omega)) is used}

\item{...}{extra parameters to pass to the optim function}
}
\value{
A tdmorefit_mixture object
}
\description{
Calculate the Empirical Bayesian Estimate parameters that predict
a given dataset using a mixture model.
}
