% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSet.R
\name{findFirstCompatibleModel}
\alias{findFirstCompatibleModel}
\title{Find the first model in the TDMore set compatible with the specified covariates.}
\usage{
findFirstCompatibleModel(tdmore_set, covariates)
}
\arguments{
\item{tdmore_set}{a tdmore_set model}

\item{covariates}{the model covariates}
}
\value{
A tdmore or tdmore_mixture model
}
\description{
Find the first model in the TDMore set compatible with the specified covariates.
}
