% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{formula.tdmore}
\alias{formula.tdmore}
\title{Get the original model}
\usage{
\method{formula}{tdmore}(x, ...)
}
\arguments{
\item{x}{TDMore object}

\item{...}{ignored}
}
\value{
The original model used when defining the TDMore object
}
\description{
Get the original model
}
