% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{formulation}
\alias{formulation}
\title{Create a new formulation object.}
\usage{
formulation(
  name,
  unit,
  dosing_interval = 24,
  default_value = 1,
  round_function = function(dose) {     dose }
)
}
\arguments{
\item{name}{the formulation name}

\item{unit}{the unit}

\item{dosing_interval}{the dosing interval in hours, 24h is the default value}

\item{default_value}{the default dose value, same unit as the one provided}

\item{round_function}{the rounding function to round the recommended dose, by default doses are not rounded}
}
\value{
a tdmore_formulation object
}
\description{
Create a new formulation object.
}
\seealso{
Other metadata: 
\code{\link{covariate}()},
\code{\link{getDosingInterval}()},
\code{\link{getMetadataByClass}()},
\code{\link{getMetadataByName}()},
\code{\link{metadata}()},
\code{\link{output}()},
\code{\link{target}()}
}
\concept{metadata}
