% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{getFixedParametersIndexes}
\alias{getFixedParametersIndexes}
\title{Get the indexes of the fixed parameters.}
\usage{
getFixedParametersIndexes(parNames, fix)
}
\arguments{
\item{parNames}{parameter names in the right order, possibly with repeated values}

\item{fix}{named numeric vector with the fixed parameters (user input)}
}
\value{
the fixed parameters indexes
}
\description{
Get the indexes of the fixed parameters.
}
\keyword{internal}
