% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{getMetadataByClass}
\alias{getMetadataByClass}
\title{Get metadata by class.}
\usage{
getMetadataByClass(tdmore, metaClass, all = FALSE)
}
\arguments{
\item{tdmore}{the tdmore model}

\item{metaClass}{the class of the metadata}

\item{all}{whether to return only a single element, or all of them (in a list)}
}
\value{
the metadata items that matches the given class
}
\description{
Get metadata by class.
}
\seealso{
Other metadata: 
\code{\link{covariate}()},
\code{\link{formulation}()},
\code{\link{getDosingInterval}()},
\code{\link{getMetadataByName}()},
\code{\link{metadata}()},
\code{\link{output}()},
\code{\link{target}()}
}
\concept{metadata}
