% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{getMetadataByName}
\alias{getMetadataByName}
\title{Get metadata (output, formulation or covariate) by name.}
\usage{
getMetadataByName(tdmore, metaName, all = FALSE)
}
\arguments{
\item{tdmore}{the tdmore model}

\item{metaName}{the name of the metadata}

\item{all}{whether to return only a single element, or all of them (in a list)}
}
\value{
the first metadata item that matches the given name
}
\description{
Get metadata (output, formulation or covariate) by name.
}
\seealso{
Other metadata: 
\code{\link{covariate}()},
\code{\link{formulation}()},
\code{\link{getDosingInterval}()},
\code{\link{getMetadataByClass}()},
\code{\link{metadata}()},
\code{\link{output}()},
\code{\link{target}()}
}
\concept{metadata}
