% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.qdoptimize.R
\name{getTroughs}
\alias{getTroughs}
\title{Automatically guesses the target troughs for a given regimen}
\usage{
getTroughs(model, regimen, deltamin = 1/4, deltaplus = 1/4, adj = 1e-15)
}
\arguments{
\item{model}{tdmore model with formulation metadata}

\item{regimen}{a treatment regimen}

\item{deltamin}{how much can we move the trough back to match an existing treatment time? in percentage of interdose-interval}

\item{deltaplus}{how much can we move the trough forward to match an existing treatment time? in percentage of interdose-interval
This method emits an error if any other treatment is found in the interval between `time` and `time+ii*(1-deltamin)`}

\item{adj}{some prediction models will display the peak rather than the trough if we use the exact treatment time.
To counteract this, `trough - adj` is returned.}
}
\value{
a numeric vector with the corresponding troughs
}
\description{
Automatically guesses the target troughs for a given regimen
}
