% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ipred}
\alias{ipred}
\title{Create a population prediction.}
\usage{
ipred(mapping = NULL, data = NULL, geom = "line",
  position = "identity", color = "steelblue2", ..., newdata = NULL,
  regimen = NULL, parameters = NULL, covariates = NULL,
  show.legend = NA, inherit.aes = TRUE)
}
\description{
Create a population prediction.
}
