% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{logLik.tdmorefit}
\alias{logLik.tdmorefit}
\title{Get the log-likelihood of the predicted values,
or try to calculate it}
\usage{
\method{logLik}{tdmorefit}(object, ...)
}
\arguments{
\item{object}{A tdmorefit object}

\item{...}{ignored}
}
\value{
A numeric value
}
\description{
Get the log-likelihood of the predicted values,
or try to calculate it
}
