% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{meltPredictions}
\alias{meltPredictions}
\title{Melt prediction results.}
\usage{
meltPredictions(x, se = FALSE)
}
\arguments{
\item{x}{dataframe returned by the predict function}

\item{se}{standard error in input dataframe, logical value}
}
\value{
the melted dataframe
}
\description{
Melt prediction results.
}
