% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{Append metadata to a TDMore model}
\usage{
metadata(tdmore, ...)
}
\arguments{
\item{tdmore}{a tdmore_output object}

\item{...}{metadata}
}
\value{
a tdmore model
}
\description{
Append metadata to a TDMore model
}
\seealso{
Other metadata: 
\code{\link{covariate}()},
\code{\link{formulation}()},
\code{\link{getDosingInterval}()},
\code{\link{getMetadataByClass}()},
\code{\link{getMetadataByName}()},
\code{\link{output}()},
\code{\link{target}()}
}
\concept{metadata}
