% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{model.frame.tdmorefit}
\alias{model.frame.tdmorefit}
\title{Get the observed values used to provide this model fit.}
\usage{
\method{model.frame}{tdmorefit}(
  formula,
  data = NULL,
  se = FALSE,
  level = 0.95,
  onlyOutput = FALSE,
  ...
)
}
\arguments{
\item{formula}{A tdmorefit object}

\item{data}{Data.frame to append and modify. If NULL, uses the observed values.}

\item{se}{TRUE to generate an additional xx.upper and xx.lower column for every xx column in the data dataset}

\item{level}{numeric value to specify the confidence interval}

\item{onlyOutput}{only keep TIME and output variables in the data.frame}

\item{...}{ignored}
}
\value{
A data.frame, similar to the one used to estimate this `tdmorefit` object.
If `se` was specified, then a column xx.upper and xx.lower with the
lower and upper confidence interval (based on the residual error model) is added.
}
\description{
Get the observed values used to provide this model fit.
}
