% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraic.R
\name{model_predict.algebraic}
\alias{model_predict.algebraic}
\title{Predict for algebraic models}
\usage{
\method{model_predict}{algebraic}(model, times, regimen = data.frame(TIME
  = numeric(), AMT = numeric(), OCC = numeric()), parameters = numeric(),
  covariates = NULL, iov = NULL, extraArguments = list())
}
\arguments{
\item{model}{the algebraic model}

\item{regimen}{dataframe with column 'TIME' and adhering to standard NONMEM specifications otherwise (columns AMT, RATE, CMT)}

\item{parameters}{a named numeric vector}

\item{covariates}{named numeric vector, or data.frame with column 'TIME', and at least TIME 0}

\item{iov}{character array with the IOV terms, NULL if no IOV, IOV on KA only}

\item{newdata}{dataframe with a column 'TIME' and 'CONC'. Other columns are ignored.}
}
\value{
A data.frame similar to the newdata data frame, but with CONC column filled out.
}
\description{
We only support dosing into the default compartment, and only bolus doses
}
\keyword{internal}
