% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_predict}
\alias{model_predict}
\title{Prototype predict function}
\usage{
model_predict(model, newdata, regimen, parameters, covariates,
  extraArguments)
}
\arguments{
\item{model}{The model itself.}

\item{newdata}{dataframe with at least a column 'TIME' and other values. A prediction will be generated for each value.}

\item{regimen}{dataframe with column 'TIME' and adhering to standard NONMEM specifications otherwise (columns AMT, RATE, CMT)}

\item{parameters}{named vector}

\item{covariates}{named vector, or data.frame with column 'TIME', and at least TIME 0}

\item{extraArguments}{named list with extra arguments to use for call}
}
\value{
A data.frame similar to the observed data frame, but with predicted values.
}
\description{
Prototype predict function
}
