% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deSolve.R
\name{model_predict.tdmore_deSolve}
\alias{model_predict.tdmore_deSolve}
\title{Predict for tdmore_deSolve}
\usage{
\method{model_predict}{tdmore_deSolve}(model, times,
  regimen = data.frame(TIME = c()), parameters = c(),
  covariates = NULL, iov = NULL, extraArguments = list())
}
\arguments{
\item{model}{The model function}

\item{times}{sampling times at which to generate a prediction}

\item{regimen}{dataframe with column 'TIME' and adhering to standard NONMEM specifications otherwise (columns AMT, RATE, CMT).}

\item{parameters}{either a dataframe with column 'TIME' and a column for each covariate and parameter, or a named numeric vector}

\item{covariates}{NOT IMPLEMENTED}

\item{iov}{character array with the IOV terms, NULL if no IOV - NOT IMPLEMENTED}

\item{extraArguments}{extra arguments to use}
}
\value{
A data.frame similar to the observed data frame, but with predicted values.
}
\description{
Predict for tdmore_deSolve
}
\keyword{internal}
