% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc.R
\name{mpc}
\alias{mpc}
\title{MPC is a generic function to make a tdmore model compatible with MPC.}
\usage{
mpc(x, parameters, ...)
}
\arguments{
\item{x}{a model object (type not defined)}

\item{parameters}{select which parameters should be mpc parameters. MPC parameters are cumulatively summed across occasions.}

\item{...}{extra arguments}
}
\value{
an object of class tdmore_mpc
}
\description{
MPC is a generic function to make a tdmore model compatible with MPC.
}
