% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc.R
\name{mpc.tdmore}
\alias{mpc.tdmore}
\title{Build a 'tdmore_mpc' object based on a 'tdmore' object.}
\usage{
\method{mpc}{tdmore}(x, theta, suffix, ...)
}
\arguments{
\item{x}{a model object (type not defined)}

\item{theta}{named vector with the initial values}

\item{suffix}{a suffix to add to the theta names}

\item{...}{extra arguments}
}
\value{
an object of class tdmore_mpc
}
\description{
Build a 'tdmore_mpc' object based on a 'tdmore' object.
}
