% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{observed_variables}
\alias{observed_variables}
\title{Create an 'observed variables' object.
This metadata indicates that these variables are interesting outputs of the model,
and should be included in a plot.}
\usage{
observed_variables(variables)
}
\arguments{
\item{variables}{variables to be observed}
}
\value{
a tdmore_observed_variables object
}
\description{
Create an 'observed variables' object.
This metadata indicates that these variables are interesting outputs of the model,
and should be included in a plot.
}
