% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.qdoptimize.R
\name{optimize}
\alias{optimize}
\title{Optimize the regimen, using the metadata defined by the model}
\usage{
optimize(fit, regimen = fit$regimen, targetMetadata = NULL)
}
\arguments{
\item{fit}{tdmorefit object}

\item{regimen}{the treatment regimen to optimize}

\item{targetMetadata}{defined targets as list(min=X, max=Y). If NULL, taken from the model metadata.}
}
\description{
This performs a step-wise optimization.
}
