% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{output}
\alias{output}
\title{Create a new output object.}
\usage{
output(name, label, unit = NULL, default_value = 1)
}
\arguments{
\item{name}{the output name}

\item{label}{the output label}

\item{unit}{the unit, can be null}

\item{default_value}{the default dose value, same unit as the one provided}
}
\value{
a tdmore_output object
}
\description{
Create a new output object.
}
\seealso{
Other metadata: 
\code{\link{covariate}()},
\code{\link{formulation}()},
\code{\link{getDosingInterval}()},
\code{\link{getMetadataByClass}()},
\code{\link{getMetadataByName}()},
\code{\link{metadata}()},
\code{\link{target}()}
}
\concept{metadata}
