% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraic.R
\name{pk1cptivbolusVCL}
\alias{pk1cptivbolusVCL}
\title{A function to predict concentration in a 1cpt PK model with bolus administration}
\usage{
pk1cptivbolusVCL(THETA = list(V = 10, CL = 5), OMEGA = list(V = 0.2, CL
  = 0.3))
}
\arguments{
\item{THETA}{a list(V=xx, CL=yy) with all typical values}

\item{OMEGA}{a list(V=0.20, CL=0.30) with all variances describing IIV}
}
\value{
the algebraic definition to predict PK 1cpt concentration, useable in algebraic()
}
\description{
A function to predict concentration in a 1cpt PK model with bolus administration
}
\examples{
predictFunction <- pk1cptivbolusVCL(
  THETA=list(V=10, CL=5),
  OMEGA=list(V=0.20, CL=0.30))
model <- algebraic(predictFunction)
tdmore <- tdmore(model, res_var=list(errorModel("CONC", prop=0.10)))
}
