% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraic.R
\name{pk1cptoralbolusVCL}
\alias{pk1cptoralbolusVCL}
\title{A function to predict concentration in a 1cpt PK model with bolus administration}
\usage{
pk1cptoralbolusVCL(THETA = list(KA = 0.5, V = 10, CL = 5),
  OMEGA = list(KA = 0, V = 0.2, CL = 0.3))
}
\arguments{
\item{THETA}{a list(KA=xx, V=xx, CL=yy) with all typical values}

\item{OMEGA}{a list(KA=xx, V=0.20, CL=0.30) with all variances describing IIV}
}
\value{
the algebraic definition to predict PK 1cpt concentration, useable in algebraic()
}
\description{
A function to predict concentration in a 1cpt PK model with bolus administration
}
\examples{
predictFunction <- pk1cptoralbolusVCL(
  THETA=list(KA=0.5, V=10, CL=5),
  OMEGA=list(KA=0.70, V=0.20, CL=0.30))
model <- algebraic(predictFunction)
tdmore <- tdmore(model)
}
