% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebraicLibrary.R
\name{pk_}
\alias{pk_}
\alias{pk1cptiv_}
\alias{pk1cptinfusion_}
\alias{pk1cptoral_}
\alias{pk2cptiv_}
\alias{pk2cptinfusion_}
\alias{pk2cptoral_}
\alias{pk3cptiv_}
\alias{pk3cptinfusion_}
\alias{pk3cptoral_}
\title{Algebraic equations for PK models}
\usage{
pk1cptiv_(t, A1 = 0, TIME, AMT, V, K, SS = 0, II = Inf)

pk1cptinfusion_(t, A1 = 0, TIME, AMT, RATE, V, K, SS = 0, II = Inf)

pk1cptoral_(t, A0 = 0, A1 = 0, TIME, AMT, V, K, KA, SS = 0, II = Inf)

pk2cptiv_(t, A1 = 0, A2 = 0, TIME, AMT, V, K, K12, K21, SS = 0, II = Inf)

pk2cptinfusion_(
  t,
  A1 = 0,
  A2 = 0,
  TIME,
  AMT,
  RATE,
  V,
  K,
  K12,
  K21,
  SS = 0,
  II = Inf
)

pk2cptoral_(
  t,
  A0 = 0,
  A1 = 0,
  A2 = 0,
  TIME,
  AMT,
  V,
  K,
  KA,
  K12,
  K21,
  SS = 0,
  II = Inf
)

pk3cptiv_(
  t,
  A1 = 0,
  A2 = 0,
  A3 = 0,
  TIME,
  AMT,
  V,
  K,
  K12,
  K21,
  K13,
  K31,
  SS = 0,
  II = Inf
)

pk3cptinfusion_(
  t,
  A1 = 0,
  A2 = 0,
  A3 = 0,
  TIME,
  AMT,
  RATE,
  V,
  K,
  K12,
  K21,
  K13,
  K31,
  SS = 0,
  II = Inf
)

pk3cptoral_(
  t,
  A0 = 0,
  A1 = 0,
  A2 = 0,
  A3 = 0,
  TIME,
  AMT,
  V,
  K,
  KA,
  K12,
  K21,
  K13,
  K31,
  SS = 0,
  II = Inf
)
}
\arguments{
\item{t}{sampling times}

\item{A1}{initial state for A1 at TIME}

\item{TIME}{dosing time}

\item{AMT}{dosing amount}

\item{V}{volume of distribution}

\item{K}{elimination constant}

\item{SS}{0 for no steady-state, 1 for steady-state}

\item{II}{inter-dose interval, required for steady-state calculations}

\item{RATE}{infusion rate}

\item{A0}{initial state for A0 at TIME}

\item{KA}{absorption constant}

\item{A2}{initial state for A2 at TIME}

\item{K12}{transfer rate from central compartment to peripheral compartment A}

\item{K21}{transfer rate from peripheral compartment A to central compartment}

\item{A3}{initial state for A3 at TIME}

\item{K13}{transfer rate from central compartment to peripheral compartment B}

\item{K31}{transfer rate from peripheral compartment B to central compartment}
}
\value{
vector of same size as `t`, with the concentrations
}
\description{
Algebraic equations for PK models
}
