% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tdmore}
\alias{plot.tdmore}
\title{Plot a tdmore object.}
\usage{
\method{plot}{tdmore}(x, regimen, covariates = NULL, newdata = NULL,
  bsv = TRUE, mc.subjects = 100, .progress = "none", ...)
}
\arguments{
\item{x}{the tdmore object}

\item{regimen}{the regimen to be predicted}

\item{covariates}{covariates}

\item{newdata}{a data.frame with at least TIME and any other columns to plot, NULL to plot all columns from the original observed data between time 0 and max(observationTime) or a numeric vector of times}

\item{bsv}{add between subject variability}

\item{.progress}{either "none" or "text" to see calculation progress of monte carlo simulations}

\item{...}{ignored}

\item{mc.maxpts}{number of subjects in the monte carlo simulation}
}
\value{
a ggplot object with the fitted individual curve, the 95% ci of this curve, the population prediction, the observed data and the residual error around the observed data
}
\description{
Plot a tdmore object.
}
