% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tdmore}
\alias{plot.tdmore}
\title{Plot a tdmore object.}
\usage{
\method{plot}{tdmore}(x, regimen, covariates = NULL, newdata = NULL,
  se = TRUE, mc.maxpts = 100, .progress = "none", ...)
}
\arguments{
\item{x}{the tdmore object}

\item{regimen}{the regimen to be predicted}

\item{covariates}{covariates}

\item{newdata}{a data.frame with at least TIME and any other columns to plot, NULL to plot all columns from the original observed data between time 0 and max(observationTime) or a numeric vector of times}

\item{se}{add a curve for the confidence interval around the population prediction}

\item{mc.maxpts}{maximum number of points to use for the monte carlo fit}

\item{.progress}{either "none" or "text" to see calculation progress of monte carlo simulations}

\item{...}{ignored}
}
\value{
a ggplot object with the fitted individual curve, the 95% ci of this curve, the population prediction, the observed data and the residual error around the observed data
}
\description{
Plot a tdmore object.
}
