% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{pop_ll}
\alias{pop_ll}
\title{Calculate the population log likelihood.}
\usage{
pop_ll(estimate, tdmore, observed, regimen, covariates)
}
\arguments{
\item{estimate}{the current estimate of the parameters}

\item{tdmore}{the tdmore object}

\item{observed}{the observed data, not used in the population log likelihoold}

\item{regimen}{data frame describing the treatment regimen}

\item{covariates}{the model covariates, not used in the population log likelihoold}
}
\value{
the population log likelihood
}
\description{
Calculate the population log likelihood.
}
