% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc.R
\name{predict.tdmore_mpc}
\alias{predict.tdmore_mpc}
\title{MPC predict method. This simply generates a cumulative sum for all}
\usage{
\method{predict}{tdmore_mpc}(
  object,
  newdata,
  regimen = NULL,
  parameters = NULL,
  covariates = NULL,
  se = FALSE,
  level = 0.95,
  ...
)
}
\arguments{
\item{object}{Object of class inheriting from `tdmore`}

\item{newdata}{Data.frame of new data with at least a TIME column, and additional columns.
Any columns that match the model output will be replaced by the predicted values.
Alternatively, a numeric vector used as TIME. The resulting data.frame will
contain all values predicted by the model.}

\item{regimen}{Treatment regimen, or NULL for no treatment}

\item{parameters}{A named numeric vector with the parameter values to use.
Any model parameters that are not provided are assumed `0`.}

\item{covariates}{a named numeric vector with the covariates to use.
Alternatively, a data.frame with column `TIME` and columns with time-varying covariates.
The first row of the data.frame should be time 0.}

\item{se}{TRUE to generate an additional xx.upper and xx.lower column for every xx column in the data dataset}

\item{level}{numeric value to specify the confidence interval}

\item{...}{extra arguments for the call to the model}
}
\value{
a data.frame with all observed values at the given time points
}
\description{
MPC predict method. This simply generates a cumulative sum for all
}
