% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{residuals.tdmore}
\alias{residuals.tdmore}
\title{Get the residual values of a predicted value vs the observed value}
\usage{
\method{residuals}{tdmore}(object, predicted, observed, weighted = FALSE, inverse = FALSE, ...)
}
\arguments{
\item{object}{A TDMore object}

\item{predicted}{What was predicted, as data.frame}

\item{observed}{The observed values, as data.frame}

\item{weighted}{Should the residuals be divided by the standard deviation of the error model?}

\item{inverse}{if TRUE, perform the inverse operation (assume `predicted` has IPRED values, and `observed` has IRES/IWRES values)}

\item{...}{ignored}
}
\value{
The `predicted` data.frame, with all output values replaced by IRES or IWRES
}
\description{
Get the residual values of a predicted value vs the observed value
}
