% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{residuals.tdmore}
\alias{residuals.tdmore}
\title{Get the residual values of a predicted value vs the observed value
This is based on the log-likelihood of the observed value in the residual error model}
\usage{
\method{residuals}{tdmore}(object, observed, predicted, log = TRUE, ...)
}
\arguments{
\item{object}{A TDMore object}

\item{observed}{The observed values, as data.frame}

\item{predicted}{What was predicted, as data.frame}

\item{log}{if TRUE, report as logPdf}

\item{...}{ignored}
}
\value{
A numeric vector
}
\description{
Get the residual values of a predicted value vs the observed value
This is based on the log-likelihood of the observed value in the residual error model
}
