% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{sampleMC_metrop}
\alias{sampleMC_metrop}
\title{Generate parameter samples using the Metropolis MCMC method.}
\usage{
sampleMC_metrop(
  tdmorefit,
  fix = tdmorefit$fix,
  mc.maxpts = 100,
  mc.batch = min(10, floor(mc.maxpts/2)),
  verbose = 0,
  tune = 1,
  .progress = interactive()
)
}
\arguments{
\item{tdmorefit}{a tdmorefit object}

\item{fix}{named numeric vector with the fixed parameters}

\item{mc.maxpts}{number of Monte-Carlo samples}

\item{mc.batch}{number of points to run per batch}

\item{verbose}{A switch which determines whether or not the progress of the
sampler is printed to the screen.  If \code{verbose} is greater than 0 the
iteration number, the \eqn{\theta} vector, the function value, and
the Metropolis acceptance rate are sent to the screen every \code{verbose}th
iteration.}

\item{tune}{tuning parameter to reduce `omega`}

\item{.progress}{Allows to specify a plyr-like progress object
A plyr progress object is a list with 3 function definitions: `init(N)`, `step()` and `term()`.
This can also be specified as a boolean. TRUE uses the default dplyr progress_estimated.}
}
\value{
the Monte-Carlo matrix, first column is always the 'sample' column
}
\description{
Generate parameter samples using the Metropolis MCMC method.
}
