% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyProfile.R
\name{shinyProfile}
\alias{shinyProfile}
\title{Open a Shiny Gadget to explore the log-likelihood profile
of a specific fit}
\usage{
shinyProfile(fitted, fix = NULL, ...)
}
\arguments{
\item{fitted}{A tdmorefit object}

\item{fix}{Which parameters to fix? Named vector of the parameters that are fixed and should not be profiled
Specify a FIX of NA to estimate the optimal value at that specific grid-point.}

\item{...}{Extra arguments passed to `profile`}
}
\description{
Open a Shiny Gadget to explore the log-likelihood profile
of a specific fit
}
