% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebe.R
\name{summary.tdmorefit}
\alias{summary.tdmorefit}
\title{Summary of a tdmorefit object.}
\usage{
\method{summary}{tdmorefit}(object, ...)
}
\arguments{
\item{object}{a tdmorefit object}

\item{...}{additional parameters ignored}
}
\description{
Summary of a tdmorefit object.
}
