% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R, R/model.R, R/nlmixr.R
\name{tdmore}
\alias{tdmore}
\alias{tdmore.RxODE}
\alias{tdmore.nlmixrUI}
\alias{tdmore.nlmixrFitCore}
\title{Append TDM functionality to a pharmacometrics structural model.}
\usage{
\method{tdmore}{RxODE}(model, res_var, parameters = NULL, omega = NULL, iov = NULL, ...)

tdmore(model, ...)

\method{tdmore}{nlmixrUI}(model, iov = NULL, ...)

\method{tdmore}{nlmixrFitCore}(model, ...)
}
\arguments{
\item{model}{the base model}

\item{res_var}{the residual variability}

\item{parameters}{list of parameter names, or NULL to automatically detect
The automatic detection will analyze omega first, to see if there are names present.
If not, it will use all parameters from the RxODE model}

\item{omega}{omega variance-covariance matrix, or NULL to use a diagonal matrix of variance 1 for all input parameters}

\item{iov}{list of parameter names related to IOV, NULL if no IOV}

\item{...}{extra arguments will be passed to the underlying structural model}
}
\value{
An object of class tdmore, which can be used to estimate posthoc bayesian parameters
}
\description{
Append TDM functionality to a pharmacometrics structural model.
}
\note{
You can use a named omega parameter to distinguish between unexplained variability (described by an a priori distribution) and
a covariate (assumed to be known for all individuals). Any input parameters not provided in omega are assumed to be covariates.
}
\examples{
modelCode <- "
CL = 23.6 * exp(ETA1*0.42);
Vc = 1070 * exp(ETA2*1.11);
ka=4.48;
CONC = centr / Vc * 1000;
d/dt(abs) = -ka*abs;
d/dt(centr) = ka*abs - CL/Vc*centr;
"
rxodeModel <- RxODE::RxODE(modelCode)
model <- tdmore(rxodeModel, res_var=list(errorModel(var="CONC", add=3.7)))
}
