% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMixture.R
\name{tdmore_mixture_covariates}
\alias{tdmore_mixture_covariates}
\title{Create a TDMore mixture model, based on several options for a discrete covariate.}
\usage{
tdmore_mixture_covariates(model, probs, covariates)
}
\arguments{
\item{model}{a tdmore model with a discrete covariate}

\item{probs}{a numeric vector with the a priori probabilities for the covariate values}

\item{covariates}{a list with covariate values corresponding to the right names}
}
\value{
a tdmore_mixture object
}
\description{
Create a TDMore mixture model, based on several options for a discrete covariate.
}
\examples{
\dontrun{
tdmore_mixture_covariates(tacrolimus_storset,
          probs=c(0.5, 0.5),
          covariates=list(c(CYP3A5=1), c(CYP3A5=0))
)
}
}
