% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSet.R
\name{tdmore_set}
\alias{tdmore_set}
\title{Create a TDMore set.}
\usage{
tdmore_set(...)
}
\arguments{
\item{...}{1 or more tdmore models, to be added in the set.}
}
\value{
a tdmore_set object
}
\description{
Create a TDMore set.
}
